feat: Refactor parser tests.
The parser tests have been refactored so that we don't have a 5000 line
test file.
