+++ src/lighting/parser/tests.rs
@@ -12,5199 +12,22 @@
 // this program. If not, see <https://www.gnu.org/licenses/>.
 //
 #[cfg(test)]
-#[allow(clippy::module_inception)]
-mod tests {
-    use crate::lighting::effects::{Color, CycleDirection, EffectLayer, EffectType};
-    use crate::lighting::parser::fixture_venue::parse_fixture_definition;
-    use crate::lighting::parser::grammar::{LightingParser, Rule};
-    use crate::lighting::parser::utils::parse_time_string;
-    use crate::lighting::parser::*;
-    use crate::lighting::tempo::TempoTransition;
-    use pest::Parser;
-    use std::time::Duration;
-
-    #[test]
-    fn test_parse_multiple_shows() {
-        let content = r#"show "Show 1" {
-    @00:00.000
-    front_wash: static color: "blue", dimmer: 60%
-}
-
-show "Show 2" {
-    @00:00.000
-    back_wash: static color: "red", dimmer: 80%
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        println!("Found {} shows", shows.len());
-        assert_eq!(shows.len(), 2);
-        assert!(shows.contains_key("Show 1"));
-        assert!(shows.contains_key("Show 2"));
-    }
-
-    #[test]
-    fn test_parse_invalid_syntax() {
-        let content = r#"show "Invalid Show" {
-    @00:00.000
-    front_wash: invalid_effect color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_parse_malformed_timing() {
-        let content = r#"show "Invalid Timing" {
-    @invalid_time
-    front_wash: static color: "blue", dimmer: 60%
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_parse_empty_show() {
-        let content = r#"show "Empty Show" {
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        let show = &shows["Empty Show"];
-        assert_eq!(show.cues.len(), 0);
-    }
-
-    #[test]
-    fn test_color_parsing() {
-        use crate::lighting::effects::Color;
-
-        // Test hex colors
-        let red = Color::from_hex("#ff0000").unwrap();
-        assert_eq!(red.r, 255);
-        assert_eq!(red.g, 0);
-        assert_eq!(red.b, 0);
-
-        // Test named colors
-        let blue = Color::from_name("blue").unwrap();
-        assert_eq!(blue.r, 0);
-        assert_eq!(blue.g, 0);
-        assert_eq!(blue.b, 255);
-
-        // Test invalid hex
-        assert!(Color::from_hex("invalid").is_err());
-
-        // Test invalid color name
-        assert!(Color::from_name("invalid").is_err());
-    }
-
-    #[test]
-    fn test_time_parsing() {
-        // Test MM:SS.mmm format
-        let time1 = parse_time_string("01:30.500").unwrap();
-        assert_eq!(time1.as_millis(), 90500);
-
-        // Test SS.mmm format
-        let time2 = parse_time_string("30.500").unwrap();
-        assert_eq!(time2.as_millis(), 30500);
-
-        // Test edge cases
-        let time3 = parse_time_string("00:00.000").unwrap();
-        assert_eq!(time3.as_millis(), 0);
-    }
-
-    #[test]
-    fn test_parse_crossfade_example() {
-        let content = r#"show "Crossfade Test" {
-    @00:00.000
-    front_wash: static color: "blue", up_time: 2s, down_time: 1s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-
-        let show = shows.get("Crossfade Test").unwrap();
-        assert_eq!(show.name, "Crossfade Test");
-        assert_eq!(show.cues.len(), 1);
-
-        let cue = &show.cues[0];
-        assert_eq!(cue.time, Duration::from_millis(0));
-        assert_eq!(cue.effects.len(), 1);
-
-        let effect = &cue.effects[0];
-        assert_eq!(effect.groups, vec!["front_wash"]);
-        assert_eq!(effect.up_time, Some(Duration::from_secs(2)));
-        assert_eq!(effect.down_time, Some(Duration::from_secs(1)));
-        println!(
-            "Timing parsing works! up_time: {:?}, down_time: {:?}",
-            effect.up_time, effect.down_time
-        );
-    }
-
-    #[test]
-    fn test_parse_zero_fade() {
-        let content = r#"show "Zero Fade Test" {
-    @00:00.000
-    front_wash: static color: "blue", up_time: 0s, down_time: 0s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Zero Fade Test").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, Some(Duration::from_secs(0)));
-        assert_eq!(effect.down_time, Some(Duration::from_secs(0)));
-        println!(
-            "Zero timing parsing works! up_time: {:?}, down_time: {:?}",
-            effect.up_time, effect.down_time
-        );
-    }
-
-    #[test]
-    fn test_parse_layering_partial() {
-        let content = r#"show "Effect Layering Demo" {
-    @00:00.000
-    # Background layer: Static blue color with 2 second fade in
-    front_wash: static color: "blue", dimmer: 100%, layer: background, blend_mode: replace, up_time: 2s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Effect Layering Demo").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, Some(Duration::from_secs(2)));
-        println!(
-            "Layering partial parsing works! up_time: {:?}",
-            effect.up_time
-        );
-    }
-
-    #[test]
-    fn test_parse_layering_2lines() {
-        let content = r#"show "Effect Layering Demo" {
-    @00:00.000
-    # Background layer: Static blue color with 2 second fade in
-    front_wash: static color: "blue", dimmer: 100%, layer: background, blend_mode: replace, up_time: 2s
-    
-    @00:02.000
-    # Midground layer: Dimmer effect that slowly dims the blue with crossfades
-    front_wash: dimmer start_level: 1.0, end_level: 0.5, duration: 5s, layer: midground, blend_mode: multiply, up_time: 1s, down_time: 1s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Effect Layering Demo").unwrap();
-        assert_eq!(show.cues.len(), 2);
-        println!("Layering 2 lines parsing works!");
-    }
-
-    #[test]
-    fn test_parse_layering_3lines() {
-        let content = r#"show "Effect Layering Demo" {
-    @00:00.000
-    # Background layer: Static blue color with 2 second fade in
-    front_wash: static color: "blue", dimmer: 100%, layer: background, blend_mode: replace, up_time: 2s
-    
-    @00:02.000
-    # Midground layer: Dimmer effect that slowly dims the blue with crossfades
-    front_wash: dimmer start_level: 1.0, end_level: 0.5, duration: 5s, layer: midground, blend_mode: multiply, up_time: 1s, down_time: 1s
-    
-    @00:04.000
-    # Foreground layer: Strobe effect on top of the dimmed blue with crossfades
-    front_wash: strobe frequency: 2, layer: foreground, blend_mode: overlay, up_time: 0.5s, down_time: 0.5s, duration: 6s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Effect Layering Demo").unwrap();
-        assert_eq!(show.cues.len(), 3);
-        println!("Layering 3 lines parsing works!");
-    }
-
-    #[test]
-    fn test_parse_strobe_simple() {
-        let content = r#"show "Test" {
-    @00:00.000
-    front_wash: strobe frequency: 2, up_time: 0.5s, down_time: 0.5s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, Some(Duration::from_millis(500)));
-        assert_eq!(effect.down_time, Some(Duration::from_millis(500)));
-        println!(
-            "Strobe simple parsing works! up_time: {:?}, down_time: {:?}",
-            effect.up_time, effect.down_time
-        );
-    }
-
-    #[test]
-    fn test_parse_strobe_no_crossfade() {
-        let content = r#"show "Test" {
-    @00:00.000
-    front_wash: strobe frequency: 2
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, None);
-        assert_eq!(effect.down_time, None);
-        println!("Strobe no crossfade parsing works!");
-    }
-
-    #[test]
-    fn test_parse_strobe_crossfade_minimal() {
-        let content = r#"show "Test" {
-    @00:00.000
-    front_wash: strobe frequency: 2, up_time: 0.5s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, Some(Duration::from_millis(500)));
-        assert_eq!(effect.down_time, None);
-        println!(
-            "Strobe timing minimal parsing works! up_time: {:?}",
-            effect.up_time
-        );
-    }
-
-    #[test]
-    fn test_parse_static_crossfade() {
-        let content = r#"show "Test" {
-    @00:00.000
-    front_wash: static color: "blue", up_time: 0.5s, down_time: 0.5s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, Some(Duration::from_millis(500)));
-        assert_eq!(effect.down_time, Some(Duration::from_millis(500)));
-        println!(
-            "Static timing parsing works! up_time: {:?}, down_time: {:?}",
-            effect.up_time, effect.down_time
-        );
-    }
-
-    #[test]
-    fn test_parse_fade_in_only() {
-        let content = r#"show "Test" {
-    @00:00.000
-    front_wash: static color: "blue", up_time: 0.5s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, Some(Duration::from_millis(500)));
-        assert_eq!(effect.down_time, None);
-        println!("Up time only parsing works! up_time: {:?}", effect.up_time);
-    }
-
-    #[test]
-    fn test_parse_fade_in_simple() {
-        let content = r#"show "Test" {
-    @00:00.000
-    front_wash: static color: "blue", up_time: 2s
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-        }
-        assert!(result.is_ok());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-        let effect = &show.cues[0].effects[0];
-
-        assert_eq!(effect.up_time, Some(Duration::from_secs(2)));
-        assert_eq!(effect.down_time, None);
-        println!(
-            "Up time simple parsing works! up_time: {:?}",
-            effect.up_time
-        );
-    }
-
-    #[test]
-    fn test_parse_fixture_type() {
-        let content = r#"fixture_type "RGBW_Par" {
-            channels: 4
-            channel_map: {
-                "dimmer": 1,
-                "red": 2,
-                "green": 3,
-                "blue": 4
-            }
-            special_cases: ["RGB", "Dimmer"]
-        }"#;
-
-        let result = parse_fixture_types(content).unwrap();
-        assert_eq!(result.len(), 1);
-
-        let fixture_type = result.get("RGBW_Par").unwrap();
-        assert_eq!(fixture_type.name(), "RGBW_Par");
-        assert_eq!(fixture_type.channels().get("dimmer"), Some(&1));
-        assert_eq!(fixture_type.channels().get("red"), Some(&2));
-        assert_eq!(fixture_type.channels().get("green"), Some(&3));
-        assert_eq!(fixture_type.channels().get("blue"), Some(&4));
-        // Note: special_cases field was removed from FixtureType
-    }
-
-    #[test]
-    fn test_parse_venue() {
-        let content = r#"venue "Club Venue" { }"#;
-
-        let result = parse_venues(content).unwrap();
-        assert_eq!(result.len(), 1);
-
-        let venue = result.get("Club Venue").unwrap();
-        assert_eq!(venue.name(), "Club Venue");
-        assert_eq!(venue.fixtures().len(), 0);
-        assert_eq!(venue.groups().len(), 0);
-    }
-
-    #[test]
-    fn test_fixture_universe_address_parsing() {
-        // Test that fixture parsing correctly extracts universe and address
-        let fixture_content = r#"fixture "Block1" Astera-PixelBrick @ 1:1"#;
-
-        // First test if the grammar can parse the fixture rule
-        match LightingParser::parse(Rule::fixture, fixture_content) {
-            Ok(mut pairs) => {
-                if let Some(pair) = pairs.next() {
-                    let fixture =
-                        parse_fixture_definition(pair).expect("Failed to parse fixture definition");
-                    assert_eq!(fixture.universe(), 1, "Block1 should be on universe 1");
-                    assert_eq!(fixture.start_channel(), 1, "Block1 should be at address 1");
-                } else {
-                    panic!("No fixture pair found");
-                }
-            }
-            Err(e) => {
-                panic!("Failed to parse fixture: {:?}", e);
-            }
-        }
-    }
-
-    #[test]
-    fn test_venue_with_fixtures() {
-        let content = r#"venue "test" {
-    fixture "Block1" Astera-PixelBrick @ 1:1
-    fixture "Block2" Astera-PixelBrick @ 1:5
-}"#;
-
-        let venues = parse_venues(content).expect("Failed to parse venue with fixtures");
-        assert_eq!(venues.len(), 1);
-
-        let venue = venues.get("test").expect("test venue not found");
-        assert_eq!(venue.fixtures().len(), 2);
-
-        let block1 = venue.fixtures().get("Block1").expect("Block1 not found");
-        assert_eq!(block1.universe(), 1, "Block1 should be on universe 1");
-        assert_eq!(block1.start_channel(), 1, "Block1 should be at address 1");
-
-        let block2 = venue.fixtures().get("Block2").expect("Block2 not found");
-        assert_eq!(block2.universe(), 1, "Block2 should be on universe 1");
-        assert_eq!(block2.start_channel(), 5, "Block2 should be at address 5");
-    }
-
-    #[test]
-    fn test_venue_with_comments() {
-        // Test venue with a comment at the beginning
-        let content = r#"# The built-in venue represents the lights that come with our IEM rig.
-venue "built-in" {
-    fixture "Block1" Astera-PixelBrick @ 1:1 tags ["wash", "side"]  
-    fixture "Block2" Astera-PixelBrick @ 1:5 tags ["wash", "side"]  
-    fixture "Block3" Astera-PixelBrick @ 1:9 tags ["wash", "front"]  
-    fixture "Block4" Astera-PixelBrick @ 1:13 tags ["wash", "front"]  
-    fixture "Block5" Astera-PixelBrick @ 1:17 tags ["wash", "front"]  
-    fixture "Block6" Astera-PixelBrick @ 1:21 tags ["wash", "front"]  
-    fixture "Block7" Astera-PixelBrick @ 1:25 tags ["wash", "side"]  
-    fixture "Block8" Astera-PixelBrick @ 1:29 tags ["wash", "side"]  
-}"#;
-
-        let venues = parse_venues(content).expect("Failed to parse venues with comments");
-        assert_eq!(venues.len(), 1);
-
-        let venue = venues.get("built-in").expect("built-in venue not found");
-        assert_eq!(venue.name(), "built-in");
-        assert_eq!(venue.fixtures().len(), 8);
-
-        // Verify that fixtures have correct universe and address values
-        let block3 = venue.fixtures().get("Block3").expect("Block3 not found");
-        assert_eq!(block3.universe(), 1, "Block3 should be on universe 1");
-        assert_eq!(block3.start_channel(), 9, "Block3 should be at address 9");
-
-        let block4 = venue.fixtures().get("Block4").expect("Block4 not found");
-        assert_eq!(block4.universe(), 1, "Block4 should be on universe 1");
-        assert_eq!(block4.start_channel(), 13, "Block4 should be at address 13");
-
-        let block5 = venue.fixtures().get("Block5").expect("Block5 not found");
-        assert_eq!(block5.universe(), 1, "Block5 should be on universe 1");
-        assert_eq!(block5.start_channel(), 17, "Block5 should be at address 17");
-
-        let block6 = venue.fixtures().get("Block6").expect("Block6 not found");
-        assert_eq!(block6.universe(), 1, "Block6 should be on universe 1");
-        assert_eq!(block6.start_channel(), 21, "Block6 should be at address 21");
-    }
-
-    #[test]
-    fn test_layer_command_parsing() {
-        // Test parsing layer commands in a show
-        let content = r#"show "Layer Control Test" {
-    @00:00.000
-        front_wash: static color: "blue", layer: foreground
-
-    @00:05.000
-        release(layer: foreground, time: 2s)
-
-    @00:10.000
-        clear(layer: foreground)
-
-    @00:15.000
-        freeze(layer: background)
-
-    @00:20.000
-        unfreeze(layer: background)
-
-    @00:25.000
-        master(layer: midground, intensity: 50%, speed: 200%)
-}"#;
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Layer command parsing should succeed: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Layer Control Test").expect("Show should exist");
-
-        // Check that cues were parsed
-        assert_eq!(show.cues.len(), 6, "Should have 6 cues");
-
-        // First cue has an effect, no layer commands
-        assert_eq!(show.cues[0].effects.len(), 1);
-        assert_eq!(show.cues[0].layer_commands.len(), 0);
-
-        // Second cue: release command
-        assert_eq!(show.cues[1].effects.len(), 0);
-        assert_eq!(show.cues[1].layer_commands.len(), 1);
-        let release_cmd = &show.cues[1].layer_commands[0];
-        assert_eq!(release_cmd.command_type, LayerCommandType::Release);
-        assert_eq!(release_cmd.layer, Some(EffectLayer::Foreground));
-        assert_eq!(
-            release_cmd.fade_time,
-            Some(std::time::Duration::from_secs(2))
-        );
-
-        // Third cue: clear command
-        let clear_cmd = &show.cues[2].layer_commands[0];
-        assert_eq!(clear_cmd.command_type, LayerCommandType::Clear);
-        assert_eq!(clear_cmd.layer, Some(EffectLayer::Foreground));
-
-        // Fourth cue: freeze command
-        let freeze_cmd = &show.cues[3].layer_commands[0];
-        assert_eq!(freeze_cmd.command_type, LayerCommandType::Freeze);
-        assert_eq!(freeze_cmd.layer, Some(EffectLayer::Background));
-
-        // Fifth cue: unfreeze command
-        let unfreeze_cmd = &show.cues[4].layer_commands[0];
-        assert_eq!(unfreeze_cmd.command_type, LayerCommandType::Unfreeze);
-        assert_eq!(unfreeze_cmd.layer, Some(EffectLayer::Background));
-
-        // Sixth cue: master command
-        let master_cmd = &show.cues[5].layer_commands[0];
-        assert_eq!(master_cmd.command_type, LayerCommandType::Master);
-        assert_eq!(master_cmd.layer, Some(EffectLayer::Midground));
-        assert!((master_cmd.intensity.unwrap() - 0.5).abs() < 0.01);
-        assert!((master_cmd.speed.unwrap() - 2.0).abs() < 0.01);
-    }
-
-    #[test]
-    fn test_clear_all_layers_command() {
-        // Test parsing clear() without layer parameter (clears all layers)
-        let content = r#"show "Clear All Test" {
-    @00:00.000
-        front_wash: static color: "blue", layer: foreground
-        back_wash: static color: "red", layer: background
-
-    @00:05.000
-        clear()
-}"#;
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Clear all command parsing should succeed: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Clear All Test").expect("Show should exist");
-
-        // Check that cues were parsed
-        assert_eq!(show.cues.len(), 2, "Should have 2 cues");
-
-        // Second cue: clear all command (no layer parameter)
-        assert_eq!(show.cues[1].effects.len(), 0);
-        assert_eq!(show.cues[1].layer_commands.len(), 1);
-        let clear_cmd = &show.cues[1].layer_commands[0];
-        assert_eq!(clear_cmd.command_type, LayerCommandType::Clear);
-        assert_eq!(clear_cmd.layer, None, "Clear all should have no layer");
-    }
-
-    #[test]
-    fn test_dsl_parsing_errors() {
-        // Test invalid DSL syntax
-        let invalid_syntax = r#"show "Invalid Show" {
-    @invalid_time
-    front_wash: invalid_effect
-}"#;
-        let result = parse_light_shows(invalid_syntax);
-        assert!(result.is_err(), "Invalid syntax should fail");
-
-        // Test missing show name
-        let missing_name = r#"show {
-    @00:00.000
-    front_wash: static color: "blue"
-}"#;
-        let result = parse_light_shows(missing_name);
-        assert!(result.is_err(), "Missing show name should fail");
-
-        // Test malformed time string
-        let malformed_time = r#"show "Test Show" {
-    @invalid_time
-    front_wash: static color: "blue"
-}"#;
-        let result = parse_light_shows(malformed_time);
-        assert!(result.is_err(), "Malformed time should fail");
-
-        // Test empty content
-        let empty_content = "";
-        let result = parse_light_shows(empty_content);
-        assert!(result.is_ok(), "Empty content should be OK");
-        assert_eq!(result.unwrap().len(), 0);
-
-        // Test content that looks like a show but has no valid shows
-        let no_shows = r#"// This is a comment
-some invalid content
-not a show"#;
-        let _result = parse_light_shows(no_shows);
-        // The parser may fail on invalid content, which is acceptable
-        // We just test that it doesn't panic
-    }
-
-    #[test]
-    fn test_dsl_edge_cases() {
-        // Test empty show
-        let empty_show = r#"show "Empty Show" { }"#;
-        let result = parse_light_shows(empty_show);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        assert_eq!(shows["Empty Show"].cues.len(), 0);
-
-        // Test show with overlapping cues
-        let overlapping_cues = r#"show "Overlapping Show" {
-    @00:05.000
-    front_wash: static color: "blue", dimmer: 60%
-    
-    @00:05.000
-    back_wash: static color: "red", dimmer: 80%
-}"#;
-        let result = parse_light_shows(overlapping_cues);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        assert_eq!(shows["Overlapping Show"].cues.len(), 2);
-
-        // Test show with multiple effects in one cue
-        let multiple_effects = r#"show "Multiple Effects" {
-    @00:00.000
-    front_wash: static color: "blue", dimmer: 60%
-    back_wash: static color: "red", dimmer: 80%
-}"#;
-        let result = parse_light_shows(multiple_effects);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        assert_eq!(shows["Multiple Effects"].cues.len(), 1);
-        assert_eq!(shows["Multiple Effects"].cues[0].effects.len(), 2);
-
-        // Test show with missing parameters
-        let missing_params = r#"show "Missing Params" {
-    @00:00.000
-    front_wash: static
-}"#;
-        let result = parse_light_shows(missing_params);
-        assert!(
-            result.is_ok(),
-            "Missing parameters should be handled gracefully"
-        );
-    }
-
-    #[test]
-    fn test_dsl_performance_large_file() {
-        // Create a large DSL file with many cues
-        let mut large_content = String::new();
-        large_content.push_str(r#"show "Large Show" {"#);
-
-        for i in 0..100 {
-            let time_ms = i * 1000; // 1 second intervals
-            let minutes = time_ms / 60000;
-            let seconds = (time_ms % 60000) / 1000;
-            let milliseconds = time_ms % 1000;
-
-            large_content.push_str(&format!(
-                r#"
-    @{:02}:{:02}.{:03}
-    fixture_{}: static color: "blue", dimmer: {}%"#,
-                minutes,
-                seconds,
-                milliseconds,
-                i,
-                (i % 100)
-            ));
-        }
-
-        large_content.push_str("\n}");
-
-        // Test parsing performance
-        let start = std::time::Instant::now();
-        let result = parse_light_shows(&large_content);
-        let duration = start.elapsed();
-
-        assert!(result.is_ok(), "Large file should parse successfully");
-        assert!(
-            duration.as_millis() < 1000,
-            "Parsing should be fast (< 1 second)"
-        );
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        assert_eq!(shows["Large Show"].cues.len(), 100);
-    }
-
-    #[test]
-    fn test_whitespace_handling() {
-        // Test zero whitespace
-        let no_whitespace =
-            r#"show"Test Show"{@00:00.000 front_wash:static color:"blue",dimmer:60%}"#;
-        let result = parse_light_shows(no_whitespace);
-        assert!(
-            result.is_ok(),
-            "Failed to parse DSL with zero whitespace: {:?}",
-            result
-        );
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        let show = shows.get("Test Show").unwrap();
-        assert_eq!(show.cues.len(), 1);
-        assert_eq!(show.cues[0].effects.len(), 1);
-
-        // Test minimal whitespace (just what's needed)
-        let minimal_whitespace = r#"show "Test Show" {
-@00:00.000
-front_wash: static color: "blue", dimmer: 60%
-}"#;
-        let result = parse_light_shows(minimal_whitespace);
-        assert!(
-            result.is_ok(),
-            "Failed to parse DSL with minimal whitespace: {:?}",
-            result
-        );
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        let show = shows.get("Test Show").unwrap();
-        assert_eq!(show.cues.len(), 1);
-        assert_eq!(show.cues[0].effects.len(), 1);
-
-        // Test moderate whitespace
-        let moderate_whitespace = r#"show "Test Show" {
-    @00:00.000
-    front_wash: static color: "blue", dimmer: 60%
-}"#;
-        let result = parse_light_shows(moderate_whitespace);
-        assert!(
-            result.is_ok(),
-            "Failed to parse DSL with moderate whitespace: {:?}",
-            result
-        );
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        let show = shows.get("Test Show").unwrap();
-        assert_eq!(show.cues.len(), 1);
-        assert_eq!(show.cues[0].effects.len(), 1);
-
-        // Test excessive whitespace (this might fail due to grammar limitations)
-        let excessive_whitespace = r#"
-            show    "Test Show"    {
-                @00:00.000    
-                front_wash    :    static    
-                color    :    "blue"    ,    
-                dimmer    :    60%    
-            }
-        "#;
-        let result = parse_light_shows(excessive_whitespace);
-        // This might fail due to the grammar not handling excessive whitespace well
-        if let Ok(shows) = result {
-            assert_eq!(shows.len(), 1);
-            let show = shows.get("Test Show").unwrap();
-            assert_eq!(show.cues.len(), 1);
-            assert_eq!(show.cues[0].effects.len(), 1);
-        }
-
-        // Test mixed whitespace (tabs, spaces, newlines)
-        let mixed_whitespace = r#"show	"Test Show"	{
-	@00:00.000	
-	front_wash	:	static	
-	color	:	"blue"	,	
-	dimmer	:	60%	
-}"#;
-        let result = parse_light_shows(mixed_whitespace);
-        assert!(
-            result.is_ok(),
-            "Failed to parse DSL with mixed whitespace: {:?}",
-            result
-        );
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        let show = shows.get("Test Show").unwrap();
-        assert_eq!(show.cues.len(), 1);
-        assert_eq!(show.cues[0].effects.len(), 1);
-    }
-
-    #[test]
-    fn test_extreme_whitespace_handling() {
-        // Test with very long whitespace sequences
-        let long_whitespace = format!(
-            r#"show "Test Show" {{{}@00:00.000{}front_wash: static color: "blue", dimmer: 60%{}}}"#,
-            " ".repeat(50),
-            " ".repeat(50),
-            " ".repeat(50)
-        );
-        let result = parse_light_shows(&long_whitespace);
-        assert!(
-            result.is_ok(),
-            "Failed to parse DSL with long whitespace: {:?}",
-            result
-        );
-
-        // Test with mixed whitespace characters
-        let mixed_whitespace = r#"show	"Test Show"	{
-		@00:00.000		
-		front_wash:	static	color:	"blue",	dimmer:	60%	
-	}"#;
-        let result = parse_light_shows(mixed_whitespace);
-        assert!(
-            result.is_ok(),
-            "Failed to parse DSL with mixed whitespace: {:?}",
-            result
-        );
-
-        // Test with newlines in various places
-        let newline_whitespace = r#"show
-"Test Show"
-{
-@00:00.000
-front_wash:
-static
-color:
-"blue",
-dimmer:
-60%
-}"#;
-        let result = parse_light_shows(newline_whitespace);
-        assert!(
-            result.is_ok(),
-            "Failed to parse DSL with newline whitespace: {:?}",
-            result
-        );
-    }
-
-    #[test]
-    fn test_comprehensive_dsl_parsing() {
-        // Test a comprehensive DSL file that uses various parameter types
-        let comprehensive_dsl = r#"show "Comprehensive Light Show" {
-    @00:00.000
-    front_wash: static color: "blue", dimmer: 60%
-    
-    @00:05.000
-    back_wash: static color: "red", dimmer: 80%
-    
-    @00:10.000
-    strobe_lights: static color: "green", dimmer: 100%
-    
-    @00:15.000
-    moving_heads: static color: "white", dimmer: 50%
-    
-    @00:20.000
-    dimmer_test: static color: "yellow", dimmer: 75%
-    
-    @00:25.000
-    rainbow_effect: static color: "cyan", dimmer: 90%
-    
-    @00:30.000
-    pulse_lights: static color: "magenta", dimmer: 25%
-    
-    @00:35.000
-    color_cycle: static color: "orange", dimmer: 85%
-    
-    @00:40.000
-    complex_chase: static color: "purple", dimmer: 95%
-    
-    @00:45.000
-    strobe_variation: static color: "black", dimmer: 0%
-    
-    @00:50.000
-    down_time: static color: "white", dimmer: 100%
-}"#;
-
-        let result = parse_light_shows(comprehensive_dsl);
-        if let Err(e) = &result {
-            println!("DSL parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Comprehensive DSL should parse successfully"
-        );
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        let show = shows.get("Comprehensive Light Show").unwrap();
-        assert_eq!(show.cues.len(), 11);
-
-        // Verify that different effect types are parsed
-        let first_cue = &show.cues[0];
-        assert_eq!(first_cue.effects.len(), 1);
-        // Check that it's a static effect (we can't directly compare struct variants)
-        match &first_cue.effects[0].effect_type {
-            EffectType::Static { .. } => {} // This is what we expect
-            _ => panic!("Expected static effect"),
-        }
-
-        // Verify that parameters are parsed correctly
-        let static_effect = &first_cue.effects[0];
-        // Check that the effect type has the expected parameters
-        match &static_effect.effect_type {
-            crate::lighting::effects::EffectType::Static { parameters, .. } => {
-                assert!(
-                    parameters.contains_key("color")
-                        || parameters.contains_key("red")
-                        || parameters.contains_key("green")
-                        || parameters.contains_key("blue")
-                );
-                assert!(parameters.contains_key("dimmer"));
-            }
-            _ => panic!("Expected static effect"),
-        }
-    }
-
-    #[test]
-    fn test_parameter_population() {
-        // Test that parameters are properly populated in effect types
-        let simple_dsl = r#"show "Parameter Test" {
-    @00:00.000
-    front_wash: static color: "blue", dimmer: 60%
-    
-    @00:05.000
-    back_wash: static color: "red", dimmer: 80%
-}"#;
-
-        let result = parse_light_shows(simple_dsl);
-        assert!(result.is_ok(), "Simple DSL should parse successfully");
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-
-        let show = shows.get("Parameter Test").unwrap();
-        assert_eq!(show.cues.len(), 2);
-
-        // Test first cue
-        let first_cue = &show.cues[0];
-        assert_eq!(first_cue.effects.len(), 1);
-
-        let first_effect = &first_cue.effects[0];
-        assert_eq!(first_effect.groups, vec!["front_wash"]);
-
-        // Check that parameters are stored in the EffectType
-        match &first_effect.effect_type {
-            crate::lighting::effects::EffectType::Static { parameters, .. } => {
-                assert!(
-                    parameters.contains_key("color")
-                        || parameters.contains_key("red")
-                        || parameters.contains_key("green")
-                        || parameters.contains_key("blue")
-                );
-                assert!(parameters.contains_key("dimmer"));
-            }
-            _ => panic!("Expected static effect"),
-        }
-
-        // Check that effect type is properly populated
-        match &first_effect.effect_type {
-            EffectType::Static {
-                parameters: static_params,
-                duration,
-            } => {
-                // Check that parameters were applied to the effect type
-                assert!(static_params.contains_key("red"));
-                assert!(static_params.contains_key("green"));
-                assert!(static_params.contains_key("blue"));
-                assert!(static_params.contains_key("dimmer"));
-
-                // Check specific values
-                assert_eq!(static_params.get("dimmer"), Some(&0.6)); // 60% converted to 0.6
-                assert_eq!(static_params.get("blue"), Some(&1.0)); // Blue color should be 1.0
-                assert_eq!(static_params.get("red"), Some(&0.0)); // Red should be 0.0
-                assert_eq!(static_params.get("green"), Some(&0.0)); // Green should be 0.0
-
-                // Duration should be None for static effects without duration parameter
-                assert_eq!(*duration, None);
-            }
-            _ => panic!("Expected static effect type"),
-        }
-
-        // Test second cue
-        let second_cue = &show.cues[1];
-        let second_effect = &second_cue.effects[0];
-        assert_eq!(second_effect.groups, vec!["back_wash"]);
-        // Check that the second effect has the expected parameters
-        match &second_effect.effect_type {
-            crate::lighting::effects::EffectType::Static { parameters, .. } => {
-                assert!(
-                    parameters.contains_key("color")
-                        || parameters.contains_key("red")
-                        || parameters.contains_key("green")
-                        || parameters.contains_key("blue")
-                );
-                assert!(parameters.contains_key("dimmer"));
-            }
-            _ => panic!("Expected static effect"),
-        }
-
-        // Check that red effect was properly applied
-        match &second_effect.effect_type {
-            EffectType::Static {
-                parameters: static_params,
-                ..
-            } => {
-                assert_eq!(static_params.get("dimmer"), Some(&0.8)); // 80% converted to 0.8
-                assert_eq!(static_params.get("red"), Some(&1.0)); // Red should be 1.0
-                assert_eq!(static_params.get("blue"), Some(&0.0)); // Blue should be 0.0
-                assert_eq!(static_params.get("green"), Some(&0.0)); // Green should be 0.0
-            }
-            _ => panic!("Expected static effect type"),
-        }
-    }
-
-    #[test]
-    fn test_advanced_parameter_parsing() {
-        // Test DSL that should use advanced parameter parsing functions
-        // Using the exact syntax the grammar expects
-        let advanced_dsl = r#"show "Advanced Show" {
-    @00:00.000
-    front_wash: static color: "blue", dimmer: 60%, fade: 2s
-    
-    @00:05.000
-    back_wash: cycle speed: 1.5, direction: forward
-    
-    @00:10.000
-    strobe_lights: strobe frequency: 8, intensity: 0.8, duration: 5s
-    
-    @00:15.000
-    moving_heads: chase loop: pingpong, direction: random, transition: crossfade
-    
-    @00:20.000
-    dimmer_test: dimmer start: 0%, end: 100%, duration: 3s, curve: linear
-    
-    @00:25.000
-    rainbow_effect: rainbow speed: 2.0, direction: forward
-    
-    @00:30.000
-    pulse_lights: pulse frequency: 4, intensity: 0.6, duty: 50%
-}"#;
-
-        let result = parse_light_shows(advanced_dsl);
-        if let Err(e) = &result {
-            println!("Advanced DSL parsing error: {}", e);
-        }
-        // This might fail if the grammar doesn't support the advanced syntax
-        // but it should help us understand what's actually supported
-        println!("Advanced DSL parsing result: {:?}", result);
-    }
-
-    #[test]
-    fn test_simple_advanced_parameters() {
-        // Test with just one advanced parameter at a time to isolate issues
-        let simple_advanced = r#"show "Simple Advanced" {
-    @00:00.000
-    front_wash: static color: "blue", dimmer: 60%, fade: 2s
-}"#;
-
-        let result = parse_light_shows(simple_advanced);
-        if let Err(e) = &result {
-            println!("Simple advanced DSL parsing error: {}", e);
-        }
-        println!("Simple advanced DSL parsing result: {:?}", result);
-    }
-
-    #[test]
-    fn test_custom_color_formats() {
-        // Test all three supported color formats
-        let custom_colors_dsl = r##"show "Custom Colors Show" {
-    @00:00.000
-    front_wash: static color: "#ff0000", dimmer: 60%
-    
-    @00:05.000
-    back_wash: static color: rgb(0, 255, 0), dimmer: 80%
-    
-    @00:10.000
-    side_wash: static color: "purple", dimmer: 100%
-}"##;
-
-        let result = parse_light_shows(custom_colors_dsl);
-        if let Err(e) = &result {
-            println!("Custom colors DSL parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Custom colors DSL should parse successfully"
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Custom Colors Show").unwrap();
-        assert_eq!(show.cues.len(), 3);
-
-        // Verify that colors are parsed correctly in the effect types
-        let first_cue = &show.cues[0];
-        match &first_cue.effects[0].effect_type {
-            crate::lighting::effects::EffectType::Static { parameters, .. } => {
-                assert!(
-                    parameters.contains_key("color")
-                        || parameters.contains_key("red")
-                        || parameters.contains_key("green")
-                        || parameters.contains_key("blue")
-                );
-            }
-            _ => panic!("Expected static effect"),
-        }
-
-        let second_cue = &show.cues[1];
-        match &second_cue.effects[0].effect_type {
-            crate::lighting::effects::EffectType::Static { parameters, .. } => {
-                assert!(
-                    parameters.contains_key("color")
-                        || parameters.contains_key("red")
-                        || parameters.contains_key("green")
-                        || parameters.contains_key("blue")
-                );
-            }
-            _ => panic!("Expected static effect"),
-        }
-
-        let third_cue = &show.cues[2];
-        match &third_cue.effects[0].effect_type {
-            crate::lighting::effects::EffectType::Static { parameters, .. } => {
-                assert!(
-                    parameters.contains_key("color")
-                        || parameters.contains_key("red")
-                        || parameters.contains_key("green")
-                        || parameters.contains_key("blue")
-                );
-            }
-            _ => panic!("Expected static effect"),
-        }
-    }
-
-    #[test]
-    fn test_user_dsl_syntax() {
-        let content = r#"show "Shieldbrother" {
-@00:00.000
-front_wash: static, color: "blue"
-
-@00:05.000
-all_wash: cycle, color: "red", color: "green", color: "blue", speed: 1.5, direction: "forward"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-        assert!(shows.contains_key("Shieldbrother"));
-
-        let show = shows.get("Shieldbrother").unwrap();
-        assert_eq!(show.cues.len(), 2);
-
-        // Check first cue (static effect)
-        let first_cue = &show.cues[0];
-        assert_eq!(first_cue.time.as_nanos(), 0);
-        assert_eq!(first_cue.effects.len(), 1);
-        assert_eq!(first_cue.effects[0].groups, vec!["front_wash"]);
-
-        // Check second cue (cycle effect)
-        let second_cue = &show.cues[1];
-        assert_eq!(second_cue.time.as_secs(), 5);
-        assert_eq!(second_cue.effects.len(), 1);
-        assert_eq!(second_cue.effects[0].groups, vec!["all_wash"]);
-
-        // Verify that the cycle effect has multiple colors
-        if let EffectType::ColorCycle {
-            colors,
-            speed,
-            direction,
-            transition: _,
-        } = &second_cue.effects[0].effect_type
-        {
-            assert_eq!(colors.len(), 3, "Cycle effect should have 3 colors");
-            use crate::lighting::effects::TempoAwareSpeed;
-            assert_eq!(*speed, TempoAwareSpeed::Fixed(1.5), "Speed should be 1.5");
-            assert_eq!(
-                *direction,
-                CycleDirection::Forward,
-                "Direction should be forward"
-            );
-
-            // Check that the colors are correct
-            assert_eq!(
-                colors[0],
-                Color::new(255, 0, 0),
-                "First color should be red"
-            );
-            assert_eq!(
-                colors[1],
-                Color::new(0, 255, 0),
-                "Second color should be green"
-            );
-            assert_eq!(
-                colors[2],
-                Color::new(0, 0, 255),
-                "Third color should be blue"
-            );
-        } else {
-            panic!("Expected ColorCycle effect type");
-        }
-    }
-
-    // ========================================================================
-    // TEMPO-BASED TIMING TESTS
-    // ========================================================================
-
-    #[test]
-    fn test_invalid_measure_time_formats() {
-        // These should fail to parse
-        let invalid_cases = vec![
-            "@invalid/1",
-            "@1/invalid",
-            "@1:1", // wrong separator
-            "@/1",  // missing measure
-            "@1/",  // missing beat
-        ];
-
-        for case in invalid_cases {
-            let result = LightingParser::parse(Rule::measure_time, case);
-            assert!(
-                result.is_err(),
-                "Invalid measure time format {} should fail to parse",
-                case
-            );
-        }
-    }
-
-    #[test]
-    fn test_invalid_time_signatures() {
-        // Test that syntactically invalid time signatures fail to parse
-        // Note: Semantically invalid but syntactically valid values (like 0/4)
-        // will parse successfully and need semantic validation in the implementation
-
-        let syntactically_invalid = vec![
-            "abc/4", // non-numeric numerator
-            "4/xyz", // non-numeric denominator
-            "4",     // missing denominator
-            "/4",    // missing numerator
-            "4/",    // missing denominator with slash
-            "4:4",   // wrong separator
-            "4 / 4", // spaces (not allowed in atomic rule)
-            "-4/4",  // negative numerator (not ASCII_DIGIT)
-            "4/-4",  // negative denominator (not ASCII_DIGIT)
-        ];
-
-        for sig in syntactically_invalid {
-            let content = format!(
-                r#"tempo {{
-    start: 0.0s
-    bpm: 120
-    time_signature: {}
-}}"#,
-                sig
-            );
-
-            let result = LightingParser::parse(Rule::tempo, &content);
-            assert!(
-                result.is_err(),
-                "Syntactically invalid time signature {} should fail to parse",
-                sig
-            );
-        }
-
-        // These are syntactically valid but semantically invalid
-        // The grammar will accept them, but implementation should reject them
-        let semantically_invalid = vec![
-            "0/4", // zero numerator (valid syntax, invalid semantics)
-            "4/0", // zero denominator (valid syntax, invalid semantics)
-            "0/0", // both zero (valid syntax, invalid semantics)
-        ];
-
-        for sig in semantically_invalid {
-            let content = format!(
-                r#"tempo {{
-    start: 0.0s
-    bpm: 120
-    time_signature: {}
-}}"#,
-                sig
-            );
-
-            let result = LightingParser::parse(Rule::tempo, &content);
-            // Grammar will parse these successfully
-            assert!(
-                result.is_ok(),
-                "Semantically invalid time signature {} parses successfully (needs runtime validation)",
-                sig
-            );
-        }
-
-        println!(
-            "Note: Grammar validation is syntax-only. Semantic validation \
-             (zero/negative values) should be done in the implementation."
-        );
-    }
-
-    #[test]
-    fn test_time_signature_change_with_invalid_position() {
-        // Test that time signature change requires a valid measure position
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @invalid { time_signature: 3/4 }
-    ]
-}"#;
-
-        let result = LightingParser::parse(Rule::tempo, content);
-        assert!(
-            result.is_err(),
-            "Time signature change with invalid position should fail"
-        );
-    }
-
-    #[test]
-    fn test_combined_invalid_changes() {
-        // Test various invalid combinations
-        // This documents that the grammar accepts syntactically valid but semantically invalid values
-
-        // Negative BPM - not actually possible since grammar uses ASCII_DIGIT+
-        // which doesn't include '-'
-        let content1 = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { bpm: -120 }
-    ]
-}"#;
-        let _result1 = LightingParser::parse(Rule::tempo, content1);
-        // This will actually fail to parse since '-' is not part of ASCII_DIGIT
-
-        // Zero time signature in change - syntactically valid, semantically invalid
-        let content2 = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { time_signature: 0/4 }
-    ]
-}"#;
-        let result2 = LightingParser::parse(Rule::tempo, content2);
-        // Grammar will parse this successfully, but semantic validation should reject it
-        assert!(result2.is_ok(), "Syntactically valid but semantically invalid time signature should parse (needs runtime validation)");
-
-        // These tests document that grammar validation is separate from semantic validation
-        println!(
-            "Note: Grammar accepts syntactically valid but semantically invalid values. \
-             Implementation should add semantic validation for: zero BPM, \
-             zero numerator/denominator in time signatures."
-        );
-    }
-
-    // ========================================================================
-    // BEAT POSITION VALIDATION TESTS (SEMANTIC)
-    // ========================================================================
-    // These tests document that the grammar will parse beat positions that
-    // may be semantically invalid for the current time signature.
-    // The implementation should add runtime validation.
-
-    #[test]
-    fn test_beat_position_exceeds_time_signature() {
-        // In 4/4 time, beat 5 doesn't exist (only beats 1-4)
-        // Grammar will parse this, but implementation should validate
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Invalid Beat Position" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @2/5
-    back_wash: static color: "red"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(
-            result.is_ok(),
-            "Beat position exceeding time signature parses (needs semantic validation)"
-        );
-        println!(
-            "Note: @2/5 in 4/4 time is semantically invalid (only 4 beats per measure). \
-             Implementation should validate beat positions against time signature."
-        );
-    }
-
-    #[test]
-    fn test_beat_position_valid_after_time_signature_change() {
-        // Beat 5 is invalid in 4/4 but valid in 6/8
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { time_signature: 6/8 }
-    ]
-}
-
-show "Beat Valid After Change" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @1/4
-    back_wash: static color: "red"
-    
-    @8/1
-    front_wash: static color: "green"
-    
-    @9/5
-    back_wash: static color: "yellow"
-    
-    @9/6
-    front_wash: static color: "purple"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(result.is_ok(), "Grammar should parse successfully");
-        println!(
-            "Note: Implementation should validate:\n\
-             - @1/4 is valid in 4/4 time\n\
-             - @9/5 is valid in 6/8 time (after change at @8/1)\n\
-             - @9/6 is valid in 6/8 time"
-        );
-    }
-
-    #[test]
-    fn test_beat_position_invalid_after_time_signature_change() {
-        // Beat 4 is valid in 4/4 but invalid in 3/4
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { time_signature: 3/4 }
-    ]
-}
-
-show "Beat Invalid After Change" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @1/4
-    back_wash: static color: "red"
-    
-    @8/1
-    front_wash: static color: "green"
-    
-    @9/4
-    back_wash: static color: "yellow"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(result.is_ok(), "Grammar should parse successfully");
-        println!(
-            "Note: Implementation should validate:\n\
-             - @1/4 is valid in 4/4 time\n\
-             - @9/4 is INVALID in 3/4 time (only 3 beats per measure)\n\
-             - Should raise semantic error for @9/4"
-        );
-    }
-
-    #[test]
-    fn test_beat_zero_is_invalid() {
-        // Beat 0 doesn't exist (beats are 1-indexed)
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Beat Zero" {
-    @1/0
-    front_wash: static color: "blue"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(
-            result.is_ok(),
-            "Beat 0 parses syntactically (needs semantic validation)"
-        );
-        println!(
-            "Note: @1/0 is semantically invalid (beats are 1-indexed). \
-             Implementation should reject beat positions < 1."
-        );
-    }
-
-    #[test]
-    fn test_fractional_beat_exceeds_time_signature() {
-        // Beat 4.5 in 4/4 time means halfway between beat 4 and 5
-        // But beat 5 doesn't exist in 4/4 time
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Fractional Beat Beyond Time Sig" {
-    @1/4.5
-    front_wash: static color: "blue"
-    
-    @2/4.9
-    back_wash: static color: "red"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(result.is_ok(), "Grammar should parse successfully");
-        println!(
-            "Note: Implementation should validate fractional beats:\n\
-             - @1/4.5 is INVALID in 4/4 (would be halfway to non-existent beat 5)\n\
-             - @1/4.0 or @1/4 would be valid\n\
-             - Beat values should be in range [1, time_sig_numerator]"
-        );
-    }
-
-    #[test]
-    fn test_beat_position_edge_cases_at_boundary() {
-        // Test beat positions right at the boundary of valid beats
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Boundary Cases" {
-    @1/1.0
-    front_wash: static color: "blue"
-    
-    @1/4.0
-    back_wash: static color: "red"
-    
-    @1/4.999
-    side_wash: static color: "green"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(result.is_ok(), "Grammar should parse successfully");
-        println!(
-            "Note: Implementation should validate:\n\
-             - @1/1.0 is valid (beat 1)\n\
-             - @1/4.0 is valid (beat 4)\n\
-             - @1/4.999 is valid (almost at the end of beat 4)\n\
-             - @1/5.0 would be INVALID (beat 5 doesn't exist)"
-        );
-    }
-
-    #[test]
-    fn test_multiple_time_sig_changes_with_beat_validation() {
-        // Complex scenario with multiple time signature changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @5/1 { time_signature: 3/4 },
-        @10/1 { time_signature: 6/8 },
-        @15/1 { time_signature: 5/4 }
-    ]
-}
-
-show "Complex Time Sig Changes" {
-    @1/4
-    front_wash: static color: "blue"
-    
-    @6/3
-    back_wash: static color: "red"
-    
-    @11/6
-    side_wash: static color: "green"
-    
-    @16/5
-    top_wash: static color: "yellow"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(result.is_ok(), "Grammar should parse successfully");
-        println!(
-            "Note: Implementation should validate beat positions across time signature changes:\n\
-             - @1/4 is valid in 4/4 time\n\
-             - @6/3 is valid in 3/4 time (after change at @5/1)\n\
-             - @11/6 is valid in 6/8 time (after change at @10/1)\n\
-             - @16/5 is valid in 5/4 time (after change at @15/1)"
-        );
-    }
-
-    #[test]
-    fn test_tempo_change_position_respects_time_signature() {
-        // Tempo change at a beat position that's invalid for current time signature
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 3/4
-    changes: [
-        @8/4 { bpm: 140 }
-    ]
-}"#;
-
-        let result = LightingParser::parse(Rule::tempo, content);
-        assert!(result.is_ok(), "Grammar should parse successfully");
-        println!(
-            "Note: Implementation should validate tempo change positions:\n\
-             - @8/4 is INVALID in 3/4 time (only 3 beats per measure)\n\
-             - Tempo changes should respect the active time signature"
-        );
-    }
-
-    #[test]
-    fn test_time_signature_change_position_respects_current_time_sig() {
-        // Time signature change at a beat position that's invalid for current time signature
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 3/4
-    changes: [
-        @8/4 { time_signature: 4/4 }
-    ]
-}"#;
-
-        let result = LightingParser::parse(Rule::tempo, content);
-        assert!(result.is_ok(), "Grammar should parse successfully");
-        println!(
-            "Note: Implementation should validate time signature change positions:\n\
-             - @8/4 is INVALID in 3/4 time (only 3 beats per measure)\n\
-             - Time signature changes should occur at valid positions in the CURRENT time signature"
-        );
-    }
-
-    // ========================================================================
-    // ADDITIONAL COVERAGE TESTS
-    // ========================================================================
-
-    #[test]
-    fn test_tempo_changes_with_absolute_time() {
-        // Test tempo changes using absolute time (@MM:SS.mmm) instead of measures
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @00:30.000 { bpm: 140 },
-        @01:00.500 { bpm: 160, transition: snap },
-        @01:30.000 { time_signature: 3/4 }
-    ]
-}"#;
-
-        let result = LightingParser::parse(Rule::tempo, content);
-        if let Err(e) = &result {
-            println!("Tempo changes with absolute time parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Tempo changes with absolute time positions should parse"
-        );
-    }
-
-    #[test]
-    fn test_fractional_transition_durations() {
-        // Test fractional measure and beat transitions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { bpm: 140, transition: 2.5 },
-        @16/1 { bpm: 160, transition: 1.5m },
-        @24/1 { bpm: 180, transition: 0.25m }
-    ]
-}"#;
-
-        let result = LightingParser::parse(Rule::tempo, content);
-        if let Err(e) = &result {
-            println!("Fractional transition durations parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Fractional measure and beat transitions should parse"
-        );
-    }
-
-    #[test]
-    fn test_empty_tempo_content() {
-        // Test tempo section with no content
-        let content = r#"tempo {
-}"#;
-
-        let result = LightingParser::parse(Rule::tempo, content);
-        assert!(
-            result.is_ok(),
-            "Empty tempo section should parse (fields are optional)"
-        );
-        println!(
-            "Note: Empty tempo section is syntactically valid but semantically incomplete. \
-             Implementation should require at least bpm and time_signature."
-        );
-    }
-
-    #[test]
-    fn test_measure_zero_is_invalid() {
-        // Measure 0 doesn't exist (measures are 1-indexed)
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Measure Zero" {
-    @0/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(
-            result.is_ok(),
-            "Measure 0 parses syntactically (needs semantic validation)"
-        );
-        println!(
-            "Note: @0/1 is semantically invalid (measures are 1-indexed). \
-             Implementation should reject measure positions < 1."
-        );
-    }
-
-    #[test]
-    fn test_negative_measure_numbers() {
-        // Negative measure numbers don't make sense
-        let content = r#"show "Negative Measure" {
-    @-1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        // This should fail to parse since ASCII_DIGIT doesn't include '-'
-        assert!(
-            result.is_err(),
-            "Negative measure numbers should fail to parse"
-        );
-    }
-
-    #[test]
-    fn test_tempo_change_with_ss_mmm_format() {
-        // Test SS.mmm format (without MM:) in tempo changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @30.500 { bpm: 140 },
-        @45.0 { bpm: 160 }
-    ]
-}"#;
-
-        let result = LightingParser::parse(Rule::tempo, content);
-        if let Err(e) = &result {
-            println!("Tempo change with SS.mmm format parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Tempo changes with SS.mmm time format should parse"
-        );
-    }
-
-    // ========================================================================
-    // BEAT/MEASURE-BASED EFFECT DURATIONS
-    // ========================================================================
-
-    #[test]
-    fn test_effect_duration_in_beats() {
-        // Test effect durations specified in beats
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Beat Durations" {
-    @1/1
-    front_wash: pulse color: "blue", duration: 4beats
-    
-    @2/1
-    back_wash: static color: "red", duration: 2beats
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        if let Err(e) = &result {
-            println!("Beat durations parsing error: {}", e);
-        }
-        assert!(result.is_ok(), "Effect durations in beats should parse");
-        println!(
-            "Note: Implementation must convert beat durations to absolute time using active tempo.\n\
-             At 120 BPM: 1 beat = 0.5s, so 4beats = 2s"
-        );
-    }
-
-    #[test]
-    fn test_effect_duration_in_measures() {
-        // Test effect durations specified in measures
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Measure Durations" {
-    @1/1
-    front_wash: cycle color: "red", color: "blue", duration: 2measures, loop: loop
-    
-    @4/1
-    back_wash: static color: "green", duration: 1measures
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        if let Err(e) = &result {
-            println!("Measure durations parsing error: {}", e);
-        }
-        assert!(result.is_ok(), "Effect durations in measures should parse");
-        println!(
-            "Note: Implementation must convert measure durations to absolute time.\n\
-             At 120 BPM in 4/4: 1 measure = 2s, so 2measures = 4s"
-        );
-    }
-
-    #[test]
-    fn test_fractional_beat_durations() {
-        // Test fractional beat and measure durations
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Fractional Durations" {
-    @1/1
-    front_wash: pulse color: "blue", duration: 2.5beats
-    
-    @2/1
-    back_wash: static color: "red", duration: 1.5measures
-    
-    @4/1
-    side_wash: strobe frequency: 4, duration: 0.5beats
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        if let Err(e) = &result {
-            println!("Fractional beat/measure durations parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Fractional beat/measure durations should parse"
-        );
-    }
-
-    #[test]
-    fn test_beat_based_fade_times() {
-        // Test beat-based fade times (up_time, down_time, fade_in, fade_out)
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Beat Fades" {
-    @1/1
-    front_wash: static color: "blue", up_time: 2beats
-    
-    @3/1
-    back_wash: static color: "red", up_time: 1beats, down_time: 1beats, duration: 4beats
-    
-    @8/1
-    side_wash: pulse color: "green", duration: 2measures
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        if let Err(e) = &result {
-            println!("Beat-based fade times parsing error: {}", e);
-        }
-        assert!(result.is_ok(), "Beat-based fade times should parse");
-        println!("Note: Crossfade times in beats allow effects to fade musically in sync.");
-    }
-
-    #[test]
-    fn test_beat_durations_across_tempo_change() {
-        // Test that beat durations work across tempo changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { bpm: 160, transition: 4 }
-    ]
-}
-
-show "Beat Duration Tempo Change" {
-    @7/1
-    front_wash: pulse color: "blue", duration: 4beats
-    
-    @9/1
-    back_wash: static color: "red", duration: 4beats
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        if let Err(e) = &result {
-            println!("Beat durations across tempo change parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Beat durations across tempo changes should parse"
-        );
-        println!(
-            "Note: Implementation must handle tempo-aware durations:\n\
-             - Effect at @7/1 starts at BPM 120, lasts 4 beats\n\
-             - During this effect, tempo begins transitioning at @8/1\n\
-             - Effect at @9/1 starts after tempo is fully 160 BPM\n\
-             - Each effect's 4beats will have different absolute durations!"
-        );
-    }
-
-    #[test]
-    fn test_measure_durations_across_time_signature_change() {
-        // Test measure durations across time signature changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { time_signature: 3/4 }
-    ]
-}
-
-show "Measure Duration Time Sig Change" {
-    @1/1
-    front_wash: cycle color: "red", color: "blue", duration: 2measures, loop: loop
-    
-    @9/1
-    back_wash: cycle color: "green", color: "yellow", duration: 2measures, loop: loop
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        if let Err(e) = &result {
-            println!(
-                "Measure durations across time signature change parsing error: {}",
-                e
-            );
-        }
-        assert!(
-            result.is_ok(),
-            "Measure durations across time signature changes should parse"
-        );
-        println!(
-            "Note: Implementation must handle time signature changes:\n\
-             - At @1/1 in 4/4: 2measures = 8 beats = 4s (at 120 BPM)\n\
-             - At @9/1 in 3/4: 2measures = 6 beats = 3s (at 120 BPM)\n\
-             - Same measure count, different absolute duration!"
-        );
-    }
-
-    #[test]
-    fn test_mixed_time_units_in_effects() {
-        // Test mixing absolute time and musical time in the same show
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Mixed Time Units" {
-    @1/1
-    front_wash: pulse color: "blue", duration: 500ms
-    
-    @2/1
-    back_wash: static color: "red", duration: 2beats
-    
-    @3/1
-    side_wash: strobe frequency: 4, duration: 1measures
-    
-    @5/1
-    top_wash: static color: "green", duration: 2s
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        if let Err(e) = &result {
-            println!("Mixed time units parsing error: {}", e);
-        }
-        assert!(
-            result.is_ok(),
-            "Mixing absolute and musical time units should parse"
-        );
-        println!(
-            "Note: Implementation supports both:\n\
-             - Absolute time: ms, s (fixed duration regardless of tempo)\n\
-             - Musical time: beats, measures (duration adapts to tempo)"
-        );
-    }
-
-    #[test]
-    fn test_beat_duration_without_tempo_section() {
-        // Test that beat/measure durations require semantic validation
-        let content = r#"show "No Tempo Section" {
-    @00:00.000
-    front_wash: pulse color: "blue", duration: 4beats
-}"#;
-
-        let result = LightingParser::parse(Rule::file, content);
-        assert!(
-            result.is_ok(),
-            "Beat durations without tempo section parse syntactically"
-        );
-        println!(
-            "Note: Implementation should require tempo section when using beat/measure durations.\n\
-             Grammar allows it syntactically, but semantic validation should enforce:\n\
-             - If any effect uses beats/measures, tempo section must exist\n\
-             - Should raise helpful error: 'Beat-based duration requires tempo section'"
-        );
-    }
-
-    // ============================================
-    // END-TO-END FUNCTIONALITY TESTS
-    // These tests verify that the parsed show actually works correctly,
-    // not just that it parses.
-    // ============================================
-
-    #[test]
-    fn test_end_to_end_measure_to_time_conversion() {
-        // Test that measure-based cues convert to correct absolute times
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Measure Conversion Test" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @2/1
-    back_wash: static color: "red"
-    
-    @4/1
-    side_wash: static color: "green"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parser error: {}", e);
-            println!("Full error details: {:?}", e);
-        }
-        assert!(result.is_ok(), "Show should parse successfully");
-        let shows = result.unwrap();
-        let show = shows.get("Measure Conversion Test").unwrap();
-
-        // At 120 BPM in 4/4: 1 beat = 0.5s, 1 measure = 2s
-        // Measure 1, beat 1 = 0.0s
-        // Measure 2, beat 1 = 2.0s
-        // Measure 4, beat 1 = 6.0s
-        assert_eq!(show.cues.len(), 3);
-        assert_eq!(show.cues[0].time.as_secs_f64(), 0.0);
-        assert_eq!(show.cues[1].time.as_secs_f64(), 2.0);
-        assert_eq!(show.cues[2].time.as_secs_f64(), 6.0);
-    }
-
-    #[test]
-    fn test_end_to_end_fractional_beat_conversion() {
-        // Test fractional beat positions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Fractional Beat Test" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @1/2
-    back_wash: static color: "red"
-    
-    @1/2.5
-    side_wash: static color: "green"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Fractional Beat Test").unwrap();
-
-        // At 120 BPM: 1 beat = 0.5s
-        // Measure 1, beat 1 = 0.0s
-        // Measure 1, beat 2 = 0.5s
-        // Measure 1, beat 2.5 = 0.75s
-        assert_eq!(show.cues.len(), 3);
-        assert_eq!(show.cues[0].time.as_secs_f64(), 0.0);
-        let time1 = show.cues[1].time.as_secs_f64();
-        let time2 = show.cues[2].time.as_secs_f64();
-        println!(
-            "Fractional beat test: beat 2 = {}s (expected 0.5s), beat 2.5 = {}s (expected 0.75s)",
-            time1, time2
-        );
-        assert!((time1 - 0.5).abs() < 0.001, "Expected 0.5s, got {}s", time1);
-        assert!(
-            (time2 - 0.75).abs() < 0.001,
-            "Expected 0.75s, got {}s",
-            time2
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_beat_duration_conversion() {
-        // Test that beat durations convert correctly
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Beat Duration Test" {
-    @1/1
-    front_wash: static color: "blue", duration: 4beats
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Beat Duration Test").unwrap();
-
-        // At 120 BPM: 4 beats = 2.0s
-        let effect = &show.cues[0].effects[0];
-        assert!(effect.effect_type.get_duration().is_some());
-        let duration = effect.effect_type.get_duration().unwrap();
-        assert!(
-            (duration.as_secs_f64() - 2.0).abs() < 0.001,
-            "4 beats should be 2.0s at 120 BPM"
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_measure_duration_conversion() {
-        // Test that measure durations convert correctly
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Measure Duration Test" {
-    @1/1
-    front_wash: static color: "blue", duration: 2measures
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Measure Duration Test").unwrap();
-
-        // At 120 BPM in 4/4: 2 measures = 4.0s
-        let effect = &show.cues[0].effects[0];
-        assert!(effect.effect_type.get_duration().is_some());
-        let duration = effect.effect_type.get_duration().unwrap();
-        assert!(
-            (duration.as_secs_f64() - 4.0).abs() < 0.001,
-            "2 measures should be 4.0s at 120 BPM in 4/4"
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_tempo_change_affects_timing() {
-        // Test that tempo changes affect subsequent measure-to-time conversions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 { bpm: 60 }
-    ]
-}
-
-show "Tempo Change Test" {
-    @4/1
-    front_wash: static color: "blue"
-    
-    @8/1
-    back_wash: static color: "red"
-    
-    @12/1
-    side_wash: static color: "green"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Tempo Change Test").unwrap();
-
-        // At 120 BPM: measure 4 = 6.0s (3 measures * 4 beats * 0.5 s/beat = 6s)
-        // At 120 BPM: measure 8 = 14.0s (7 measures * 4 beats * 0.5 s/beat = 14s)
-        // At 60 BPM (starting at measure 8): measure 12 = 14.0s + 16.0s = 30.0s
-        // (4 measures * 4 beats/measure * 1.0 s/beat = 16s)
-        assert_eq!(show.cues.len(), 3);
-        assert!((show.cues[0].time.as_secs_f64() - 6.0).abs() < 0.001);
-        assert!((show.cues[1].time.as_secs_f64() - 14.0).abs() < 0.001);
-        assert!((show.cues[2].time.as_secs_f64() - 30.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_end_to_end_time_signature_change_affects_timing() {
-        // Test that time signature changes affect measure calculations
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { time_signature: 3/4 }
-    ]
-}
-
-show "Time Signature Change Test" {
-    @4/1
-    front_wash: static color: "blue"
-    
-    @5/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Time Signature Change Test").unwrap();
-
-        // At 120 BPM in 4/4: measure 4 = 6.0s
-        // At 120 BPM in 3/4: measure 5 = 6.0s + 1.5s = 7.5s
-        assert_eq!(show.cues.len(), 2);
-        let time0 = show.cues[0].time.as_secs_f64();
-        let time1 = show.cues[1].time.as_secs_f64();
-        println!("Time sig change test: measure 4 = {}s (expected 6.0s), measure 5 = {}s (expected 7.5s)", time0, time1);
-        assert!((time0 - 6.0).abs() < 0.001, "Expected 6.0s, got {}s", time0);
-        assert!((time1 - 7.5).abs() < 0.001, "Expected 7.5s, got {}s", time1);
-    }
-
-    #[test]
-    fn test_end_to_end_beat_duration_with_tempo_change() {
-        // Test that beat durations use the tempo at the cue time
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 60 }
-    ]
-}
-
-show "Beat Duration Tempo Change Test" {
-    @2/1
-    front_wash: static color: "blue", duration: 4beats
-    
-    @5/1
-    back_wash: static color: "red", duration: 4beats
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Beat Duration Tempo Change Test").unwrap();
-
-        // At 120 BPM: 4 beats = 2.0s
-        let effect1 = &show.cues[0].effects[0];
-        let duration1 = effect1.effect_type.get_duration().unwrap();
-        assert!(
-            (duration1.as_secs_f64() - 2.0).abs() < 0.001,
-            "4 beats at 120 BPM should be 2.0s"
-        );
-
-        // At 60 BPM: 4 beats = 4.0s
-        // The tempo changes to 60 BPM at @4/1, so @5/1 should use 60 BPM
-        let cue1_time = show.cues[1].time;
-        let effect2 = &show.cues[1].effects[0];
-        let duration2 = effect2.effect_type.get_duration().unwrap();
-        let actual_duration = duration2.as_secs_f64();
-        println!("Beat duration with tempo change test: cue 0 at @2/1 (time={:?}), cue 1 at @5/1 (time={:?}), duration = {}s (expected 4.0s at 60 BPM)", show.cues[0].time, cue1_time, actual_duration);
-        if let Some(tm) = &show.tempo_map {
-            let bpm_at_cue0 = tm.bpm_at_time(show.cues[0].time, 0.0);
-            let bpm_at_cue1 = tm.bpm_at_time(cue1_time, 0.0);
-            println!(
-                "BPM at cue 0 time {:?} = {}, BPM at cue 1 time {:?} = {}",
-                show.cues[0].time, bpm_at_cue0, cue1_time, bpm_at_cue1
-            );
-            println!("Tempo changes: {:?}", tm.changes);
-            for change in &tm.changes {
-                if let Some(change_time) = change.position.absolute_time() {
-                    println!(
-                        "  Change at {:?}: bpm={:?}, transition={:?}",
-                        change_time, change.bpm, change.transition
-                    );
-                }
-            }
-        }
-        assert!(
-            (actual_duration - 4.0).abs() < 0.001,
-            "4 beats at 60 BPM should be 4.0s, got {}s",
-            actual_duration
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_up_time_and_down_time_with_beats() {
-        // Test that up_time and down_time work with beats
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Beat Fade Times Test" {
-    @1/1
-    front_wash: static color: "blue", up_time: 2beats, down_time: 2beats
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Beat Fade Times Test").unwrap();
-
-        // At 120 BPM: 2 beats = 1.0s
-        let effect = &show.cues[0].effects[0];
-        assert!(effect.up_time.is_some());
-        assert!(effect.down_time.is_some());
-        assert!((effect.up_time.unwrap().as_secs_f64() - 1.0).abs() < 0.001);
-        assert!((effect.down_time.unwrap().as_secs_f64() - 1.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_end_to_end_complex_tempo_changes() {
-        // Test complex scenario with multiple tempo changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 140 },
-        @8/1 { bpm: 100 },
-        @12/1 { time_signature: 3/4 }
-    ]
-}
-
-show "Complex Tempo Test" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @4/1
-    back_wash: static color: "red"
-    
-    @8/1
-    side_wash: static color: "green"
-    
-    @12/1
-    top_wash: static color: "yellow"
-    
-    @13/1
-    bottom_wash: static color: "purple"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Complex Tempo Test").unwrap();
-
-        // Verify all cues are parsed and have correct times
-        assert_eq!(show.cues.len(), 5);
-
-        // Measure 1 at 120 BPM = 0.0s
-        assert!((show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001);
-
-        // Measure 4 at 120 BPM = 6.0s
-        assert!((show.cues[1].time.as_secs_f64() - 6.0).abs() < 0.001);
-
-        // Measure 8 at 140 BPM = 6.0s + (4 measures * 4 beats/measure * 60/140) = 6.0s + 6.857s = 12.857s
-        // (Actually, we need to calculate more carefully: measures 4-8 at 140 BPM)
-        // Let's verify it's reasonable
-        assert!(show.cues[2].time.as_secs_f64() > 12.0);
-        assert!(show.cues[2].time.as_secs_f64() < 14.0);
-
-        // Measure 12 and 13 should be after the time signature change
-        assert!(show.cues[3].time.as_secs_f64() > show.cues[2].time.as_secs_f64());
-        assert!(show.cues[4].time.as_secs_f64() > show.cues[3].time.as_secs_f64());
-    }
-
-    #[test]
-    fn test_end_to_end_non_zero_start_offset() {
-        // Test that start offset is respected
-        let content = r#"tempo {
-    start: 5.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Start Offset Test" {
-    @1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Start Offset Test").unwrap();
-
-        // Measure 1, beat 1 should be at start_offset = 5.0s
-        let actual_time = show.cues[0].time.as_secs_f64();
-        if let Some(tm) = &show.tempo_map {
-            println!(
-                "Start offset test: tempo_map.start_offset = {:?}, expected 5.0s, got {}s",
-                tm.start_offset, actual_time
-            );
-        }
-        assert!(
-            (actual_time - 5.0).abs() < 0.001,
-            "Expected 5.0s, got {}s",
-            actual_time
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_measure_notation_without_tempo_error() {
-        // Test that using measure notation without tempo section fails
-        let content = r#"show "No Tempo Test" {
-    @1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_err(),
-            "Measure notation should require tempo section"
-        );
-        if let Err(e) = result {
-            assert!(
-                e.to_string().contains("tempo"),
-                "Error should mention tempo"
-            );
-        }
-    }
-
-    #[test]
-    fn test_end_to_end_beat_duration_without_tempo_error() {
-        // Test that using beat durations without tempo section fails
-        let content = r#"show "No Tempo Duration Test" {
-    @00:00.000
-    front_wash: static color: "blue", duration: 4beats
-}"#;
-
-        let result = parse_light_shows(content);
-        if result.is_ok() {
-            println!("WARNING: Parsing succeeded, but should have failed");
-            // The grammar allows beat durations without tempo (syntactic level),
-            // but the semantic implementation should catch this at duration conversion time
-            // This is by design - semantic validation happens in the parser, not the grammar
-        }
-        assert!(
-            result.is_err(),
-            "Beat durations should require tempo section"
-        );
-        if let Err(e) = result {
-            let err_msg = e.to_string();
-            println!("Error message: {}", err_msg);
-            assert!(
-                err_msg.contains("tempo") || err_msg.contains("Beat"),
-                "Error should mention tempo or beats, got: {}",
-                err_msg
-            );
-        }
-    }
-
-    #[test]
-    fn test_end_to_end_tempo_map_is_present() {
-        // Test that tempo_map is actually stored in the show
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Tempo Map Test" {
-    @1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Tempo Map Test").unwrap();
-
-        assert!(show.tempo_map.is_some(), "Tempo map should be present");
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-        assert_eq!(tempo_map.initial_bpm, 120.0);
-        assert_eq!(tempo_map.initial_time_signature.numerator, 4);
-        assert_eq!(tempo_map.initial_time_signature.denominator, 4);
-    }
-
-    #[test]
-    fn test_end_to_end_mixed_absolute_and_measure_timing() {
-        // Test that absolute time and measure timing work together
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Mixed Timing Test" {
-    @00:00.000
-    front_wash: static color: "blue"
-    
-    @1/1
-    back_wash: static color: "red"
-    
-    @00:02.000
-    side_wash: static color: "green"
-    
-    @2/1
-    top_wash: static color: "yellow"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Mixed Timing Test").unwrap();
-
-        assert_eq!(show.cues.len(), 4);
-
-        // Absolute time @00:00.000 = 0.0s
-        assert!((show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001);
-
-        // Measure @1/1 = 0.0s (same as above)
-        assert!((show.cues[1].time.as_secs_f64() - 0.0).abs() < 0.001);
-
-        // Absolute time @00:02.000 = 2.0s
-        assert!((show.cues[2].time.as_secs_f64() - 2.0).abs() < 0.001);
-
-        // Measure @2/1 = 2.0s (same as above)
-        assert!((show.cues[3].time.as_secs_f64() - 2.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_end_to_end_gradual_tempo_transition() {
-        // Test that gradual tempo transitions are handled (snap for now, but structure should work)
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 140, transition: 4 }
-    ]
-}
-
-show "Gradual Transition Test" {
-    @4/1
-    front_wash: static color: "blue"
-    
-    @6/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Gradual Transition Test").unwrap();
-
-        // The tempo change should be parsed correctly
-        assert!(show.tempo_map.is_some());
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-        assert_eq!(tempo_map.changes.len(), 1);
-
-        // Verify the transition type is stored
-        match tempo_map.changes[0].transition {
-            TempoTransition::Beats(beats, _) => assert_eq!(beats, 4.0),
-            _ => panic!("Expected Beats transition"),
-        }
-    }
-
-    #[test]
-    fn test_end_to_end_bpm_interpolation_during_gradual_transition() {
-        // Test that bpm_at_time correctly interpolates during gradual transitions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 180, transition: 4 }
-    ]
-}
-
-show "BPM Interpolation Test" {
-    @4/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("BPM Interpolation Test").unwrap();
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-
-        // Transition starts at measure 4 (6.0s at 120 BPM)
-        let change_time = tempo_map.changes[0].position.absolute_time().unwrap();
-
-        // At start of transition: should be 120 BPM
-        let bpm_start = tempo_map.bpm_at_time(change_time, 0.0);
-        assert!(
-            (bpm_start - 120.0).abs() < 0.1,
-            "BPM at transition start should be 120"
-        );
-
-        // During transition (midway): should be interpolated (120 + (180-120)*0.5 = 150)
-        // Transition duration: 4 beats at 120 BPM = 4 * 60/120 = 2.0s
-        let mid_time = change_time + Duration::from_secs(1); // 1 second into transition
-        let bpm_mid = tempo_map.bpm_at_time(mid_time, 0.0);
-        assert!(
-            (bpm_mid - 150.0).abs() < 1.0,
-            "BPM at transition midpoint should be ~150, got {}",
-            bpm_mid
-        );
-
-        // After transition: should be 180 BPM
-        let end_time = change_time + Duration::from_secs(3); // After transition completes
-        let bpm_end = tempo_map.bpm_at_time(end_time, 0.0);
-        assert!(
-            (bpm_end - 180.0).abs() < 0.1,
-            "BPM after transition should be 180"
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_file_level_tempo_applies_to_multiple_shows() {
-        // Test that file-level tempo applies to all shows without their own tempo
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Show 1" {
-    @1/1
-    front_wash: static color: "blue"
-}
-
-show "Show 2" {
-    @2/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-
-        // Both shows should have the global tempo
-        let show1 = shows.get("Show 1").unwrap();
-        let show2 = shows.get("Show 2").unwrap();
-
-        assert!(show1.tempo_map.is_some(), "Show 1 should have tempo map");
-        assert!(show2.tempo_map.is_some(), "Show 2 should have tempo map");
-
-        // Both should have the same tempo (120 BPM)
-        assert_eq!(show1.tempo_map.as_ref().unwrap().initial_bpm, 120.0);
-        assert_eq!(show2.tempo_map.as_ref().unwrap().initial_bpm, 120.0);
-
-        // Both shows should correctly convert measure-based timing
-        assert!((show1.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001);
-        assert!((show2.cues[0].time.as_secs_f64() - 2.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_end_to_end_show_specific_tempo_overrides_global() {
-        // Test that show-specific tempo overrides global tempo
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Show With Own Tempo" {
-    tempo {
-        start: 0.0s
-        bpm: 60
-        time_signature: 4/4
-    }
-    
-    @1/1
-    front_wash: static color: "blue"
-}
-
-show "Show Using Global Tempo" {
-    @1/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        if let Err(e) = &result {
-            println!("Parse error: {}", e);
-        }
-        assert!(result.is_ok(), "Parsing should succeed");
-        let shows = result.unwrap();
-
-        let show1 = shows.get("Show With Own Tempo").unwrap();
-        let show2 = shows.get("Show Using Global Tempo").unwrap();
-
-        // Show 1 should use its own tempo (60 BPM)
-        assert_eq!(show1.tempo_map.as_ref().unwrap().initial_bpm, 60.0);
-
-        // Show 2 should use global tempo (120 BPM)
-        assert_eq!(show2.tempo_map.as_ref().unwrap().initial_bpm, 120.0);
-
-        // Measure 1/1 is always at 0.0s (plus start offset) regardless of BPM
-        // The BPM affects the duration of the measure, not its start time
-        // To verify different tempos, we can check measure 2/1:
-        // - At 60 BPM: measure 2 = 4.0s (one full measure = 4 beats * 1.0s/beat)
-        // - At 120 BPM: measure 2 = 2.0s (one full measure = 4 beats * 0.5s/beat)
-        let show1_time = show1.cues[0].time.as_secs_f64();
-        let show2_time = show2.cues[0].time.as_secs_f64();
-        assert!(
-            (show1_time - 0.0).abs() < 0.001,
-            "Show 1 measure 1/1 should be 0.0s"
-        );
-        assert!(
-            (show2_time - 0.0).abs() < 0.001,
-            "Show 2 measure 1/1 should be 0.0s"
-        );
-
-        // Verify the tempo maps are correct
-        assert_eq!(show1.tempo_map.as_ref().unwrap().initial_bpm, 60.0);
-        assert_eq!(show2.tempo_map.as_ref().unwrap().initial_bpm, 120.0);
-    }
-
-    #[test]
-    fn test_end_to_end_beat_duration_during_gradual_transition() {
-        // Test that beat durations use correct BPM during gradual transitions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 180, transition: 4 }
-    ]
-}
-
-show "Beat Duration During Transition" {
-    @4/1
-    front_wash: static color: "blue", duration: 2beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Beat Duration During Transition").unwrap();
-
-        let effect = &show.cues[0].effects[0];
-
-        // Duration should integrate through the transition curve
-        // Starting at 120 BPM, transitioning to 180 BPM over 4 beats
-        // At start (120 BPM): 4 beats = 2.0s
-        // We need 2 beats starting at the beginning of the transition
-        // Since BPM is increasing during the transition, 2 beats will take slightly less than 1.0s
-        // The exact calculation integrates through the curve: approximately 0.899s
-        let duration = effect.effect_type.get_duration().unwrap();
-        // The duration should be less than 1.0s (which would be at constant 120 BPM)
-        // and more than 0.667s (which would be at constant 180 BPM)
-        assert!(
-            duration.as_secs_f64() > 0.85 && duration.as_secs_f64() < 0.95,
-            "2 beats during transition should integrate through curve: expected ~0.899s, got {}s",
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_absolute_time_tempo_changes() {
-        // Test that tempo changes at absolute time positions work correctly
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @00:06.000 { bpm: 60 }
-    ]
-}
-
-show "Absolute Time Tempo Change" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @4/1
-    back_wash: static color: "red"
-    
-    @8/1
-    side_wash: static color: "green"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Absolute Time Tempo Change").unwrap();
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-
-        // Measure 4 at 120 BPM = 6.0s (exactly when tempo changes)
-        // Measure 8: first 6 measures at 120 BPM = 6.0s, then 2 measures at 60 BPM = 8.0s, total = 14.0s
-        assert!((show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001);
-        assert!((show.cues[1].time.as_secs_f64() - 6.0).abs() < 0.001);
-
-        // Measure 8 calculation: measures 1-6 at 120 BPM = 6.0s, measures 7-8 at 60 BPM = 8.0s, total = 14.0s
-        // Note: When tempo changes are at absolute time, the calculation becomes more complex
-        // because measure positions need to be converted to absolute time first
-        let measure8_time = show.cues[2].time.as_secs_f64();
-        println!("Measure 8 time: {}s (expected ~14.0s, but calculation may vary with absolute time tempo changes)", measure8_time);
-        // The calculation is complex with absolute time tempo changes, so we just verify it's after measure 4
-        assert!(
-            measure8_time > show.cues[1].time.as_secs_f64(),
-            "Measure 8 should be after measure 4, got {}s",
-            measure8_time
-        );
-
-        // Verify the tempo change is at the correct time
-        assert_eq!(tempo_map.changes.len(), 1);
-        let change_time = tempo_map.changes[0].position.absolute_time().unwrap();
-        assert!((change_time.as_secs_f64() - 6.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_end_to_end_duration_spanning_tempo_change() {
-        // Test that beat durations integrate through tempo changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 60 }
-    ]
-}
-
-show "Duration Spanning Change" {
-    @3/1
-    front_wash: static color: "blue", duration: 8beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Duration Spanning Change").unwrap();
-
-        // Duration starts at measure 3 (4.0s at 120 BPM)
-        // 8 beats: 4 beats at 120 BPM (measure 3) = 2.0s, then 4 beats at 60 BPM (measure 4) = 4.0s
-        // Total = 6.0s
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-
-        // Measure 3 has 4 beats at 120 BPM = 2.0s
-        // Measure 4 starts when tempo changes to 60 BPM
-        // Remaining 4 beats at 60 BPM = 4.0s
-        // Total = 6.0s
-        let expected_duration = 4.0 * 60.0 / 120.0 + 4.0 * 60.0 / 60.0; // 2.0 + 4.0 = 6.0s
-        assert!(
-            (duration.as_secs_f64() - expected_duration).abs() < 0.01,
-            "Duration should integrate through tempo change: expected ~{}s, got {}s",
-            expected_duration,
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_duration_spanning_gradual_tempo_transition() {
-        // Test that beat durations integrate through gradual tempo transitions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @1/3 { bpm: 180, transition: 4 }
-    ]
-}
-
-show "Duration Spanning Gradual Transition" {
-    @1/1
-    front_wash: static color: "blue", duration: 8beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Duration Spanning Gradual Transition").unwrap();
-
-        // Starting at measure 1/beat 1, duration of 8 beats
-        // Gradual tempo change at measure 1/beat 3 (after 2 beats) from 120 to 180 over 4 beats
-        // So: 2 beats at 120 BPM = 1.0s
-        // Then 4 beats during transition (120 -> 180 linearly)
-        // Then 2 beats at 180 BPM = 2 * 60 / 180 = ~0.667s
-        // The transition: 4 beats at average BPM (150) = 1.6s
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-
-        // Verify it integrates through the gradual transition
-        // 2 beats at 120 BPM = 1.0s
-        // 4 beats during transition (average 150 BPM) = 1.6s
-        // 2 beats at 180 BPM = ~0.667s
-        // Total = ~3.267s
-        let time_before = 2.0 * 60.0 / 120.0; // 1.0s
-        let avg_bpm_during_transition = (120.0 + 180.0) / 2.0; // 150 BPM
-        let transition_time = 4.0 * 60.0 / avg_bpm_during_transition; // ~1.6s
-        let time_after = 2.0 * 60.0 / 180.0; // ~0.667s
-        let expected_duration = time_before + transition_time + time_after;
-
-        // The actual calculation uses precise integration, so there may be small differences
-        // from the approximation using average BPM. Allow a bit more tolerance.
-        assert!(
-            (duration.as_secs_f64() - expected_duration).abs() < 0.1,
-            "Duration should integrate through gradual transition: expected ~{}s, got {}s",
-            expected_duration,
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_duration_starting_mid_transition() {
-        // Test that durations starting in the middle of a gradual transition integrate correctly
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @1/1 { bpm: 180, transition: 4 }
-    ]
-}
-
-show "Duration Mid Transition" {
-    @1/2.5
-    front_wash: static color: "blue", duration: 2beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Duration Mid Transition").unwrap();
-
-        // The effect starts at measure 1, beat 2.5
-        // At 120 BPM in 4/4: measure 1, beat 1 = 0.0s, beat 2.5 = 0.75s
-        // The tempo transition starts at measure 1, beat 1 (0.0s) and transitions from 120 to 180 over 4 beats
-        // At 120 BPM: 4 beats = 2.0s, so transition completes at 2.0s
-        // At beat 2.5 (0.75s), we're 0.75s into the 2.0s transition = 37.5% through
-        // BPM at that point: 120 + (180-120) * 0.375 = 142.5 BPM
-        // We need to calculate duration for 2 beats starting from this point
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-
-        // The duration should integrate through the remaining transition
-        // At 0.75s into transition: bpm = 142.5
-        // We need to integrate 2 beats through the curve
-        // This is a complex calculation, but we verify it's reasonable
-        // At constant 142.5 BPM: 2 beats = 2 * 60 / 142.5 = 0.842s
-        // But since BPM is increasing, it should be slightly less than this
-        // At constant 180 BPM: 2 beats = 2 * 60 / 180 = 0.667s
-        // So expected should be between 0.667s and 0.842s
-        assert!(
-            duration.as_secs_f64() > 0.6 && duration.as_secs_f64() < 0.9,
-            "Duration starting mid-transition should integrate correctly: got {}s",
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_pulse_duration_spanning_tempo_change() {
-        // Test that pulse effects with beat durations integrate through tempo changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 60 }
-    ]
-}
-
-show "Pulse Duration Spanning Change" {
-    @3/1
-    front_wash: pulse color: "blue", frequency: 2, duration: 8beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Pulse Duration Spanning Change").unwrap();
-
-        // Pulse effect starts at measure 3 (4.0s at 120 BPM)
-        // 8 beats: 4 beats at 120 BPM (measure 3) = 2.0s, then 4 beats at 60 BPM (measure 4) = 4.0s
-        // Total = 6.0s (same as static effect)
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-
-        // Measure 3 has 4 beats at 120 BPM = 2.0s
-        // Measure 4 starts when tempo changes to 60 BPM
-        // Remaining 4 beats at 60 BPM = 4.0s
-        // Total = 6.0s
-        let expected_duration = 4.0 * 60.0 / 120.0 + 4.0 * 60.0 / 60.0; // 2.0 + 4.0 = 6.0s
-        assert!(
-            (duration.as_secs_f64() - expected_duration).abs() < 0.01,
-            "Pulse duration should integrate through tempo change: expected ~{}s, got {}s",
-            expected_duration,
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_strobe_duration_spanning_tempo_change() {
-        // Test that strobe effects with beat durations integrate through tempo changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 60 }
-    ]
-}
-
-show "Strobe Duration Spanning Change" {
-    @3/1
-    front_wash: strobe frequency: 4, duration: 8beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Strobe Duration Spanning Change").unwrap();
-
-        // Strobe effect starts at measure 3 (4.0s at 120 BPM)
-        // 8 beats: 4 beats at 120 BPM (measure 3) = 2.0s, then 4 beats at 60 BPM (measure 4) = 4.0s
-        // Total = 6.0s (same as static effect)
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-
-        // Measure 3 has 4 beats at 120 BPM = 2.0s
-        // Measure 4 starts when tempo changes to 60 BPM
-        // Remaining 4 beats at 60 BPM = 4.0s
-        // Total = 6.0s
-        let expected_duration = 4.0 * 60.0 / 120.0 + 4.0 * 60.0 / 60.0; // 2.0 + 4.0 = 6.0s
-        assert!(
-            (duration.as_secs_f64() - expected_duration).abs() < 0.01,
-            "Strobe duration should integrate through tempo change: expected ~{}s, got {}s",
-            expected_duration,
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_pulse_duration_spanning_gradual_transition() {
-        // Test that pulse effects with beat durations integrate through gradual tempo transitions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @1/3 { bpm: 180, transition: 4 }
-    ]
-}
-
-show "Pulse Duration Spanning Gradual Transition" {
-    @1/1
-    front_wash: pulse color: "blue", frequency: 2, duration: 8beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows
-            .get("Pulse Duration Spanning Gradual Transition")
-            .unwrap();
-
-        // Starting at measure 1/beat 1, duration of 8 beats
-        // Gradual tempo change at measure 1/beat 3 (after 2 beats) from 120 to 180 over 4 beats
-        // So: 2 beats at 120 BPM = 1.0s
-        // Then 4 beats during transition (120 -> 180 linearly)
-        // Then 2 beats at 180 BPM = 2 * 60 / 180 = ~0.667s
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-
-        // Verify it integrates through the gradual transition
-        // 2 beats at 120 BPM = 1.0s
-        // 4 beats during transition (average 150 BPM) = 1.6s
-        // 2 beats at 180 BPM = ~0.667s
-        // Total = ~3.267s
-        let time_before = 2.0 * 60.0 / 120.0; // 1.0s
-        let avg_bpm_during_transition = (120.0 + 180.0) / 2.0; // 150 BPM
-        let transition_time = 4.0 * 60.0 / avg_bpm_during_transition; // ~1.6s
-        let time_after = 2.0 * 60.0 / 180.0; // ~0.667s
-        let expected_duration = time_before + transition_time + time_after;
-
-        // The actual calculation uses precise integration, so there may be small differences
-        // from the approximation using average BPM. Allow a bit more tolerance.
-        assert!(
-            (duration.as_secs_f64() - expected_duration).abs() < 0.1,
-            "Pulse duration should integrate through gradual transition: expected ~{}s, got {}s",
-            expected_duration,
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_strobe_duration_spanning_gradual_transition() {
-        // Test that strobe effects with beat durations integrate through gradual tempo transitions
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @1/3 { bpm: 180, transition: 4 }
-    ]
-}
-
-show "Strobe Duration Spanning Gradual Transition" {
-    @1/1
-    front_wash: strobe frequency: 4, duration: 8beats
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows
-            .get("Strobe Duration Spanning Gradual Transition")
-            .unwrap();
-
-        // Starting at measure 1/beat 1, duration of 8 beats
-        // Gradual tempo change at measure 1/beat 3 (after 2 beats) from 120 to 180 over 4 beats
-        // So: 2 beats at 120 BPM = 1.0s
-        // Then 4 beats during transition (120 -> 180 linearly)
-        // Then 2 beats at 180 BPM = 2 * 60 / 180 = ~0.667s
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-
-        // Verify it integrates through the gradual transition
-        // 2 beats at 120 BPM = 1.0s
-        // 4 beats during transition (average 150 BPM) = 1.6s
-        // 2 beats at 180 BPM = ~0.667s
-        // Total = ~3.267s
-        let time_before = 2.0 * 60.0 / 120.0; // 1.0s
-        let avg_bpm_during_transition = (120.0 + 180.0) / 2.0; // 150 BPM
-        let transition_time = 4.0 * 60.0 / avg_bpm_during_transition; // ~1.6s
-        let time_after = 2.0 * 60.0 / 180.0; // ~0.667s
-        let expected_duration = time_before + transition_time + time_after;
-
-        // The actual calculation uses precise integration, so there may be small differences
-        // from the approximation using average BPM. Allow a bit more tolerance.
-        assert!(
-            (duration.as_secs_f64() - expected_duration).abs() < 0.1,
-            "Strobe duration should integrate through gradual transition: expected ~{}s, got {}s",
-            expected_duration,
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_measure_based_transition() {
-        // Test that measure-based transitions work correctly (not just beat-based)
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 180, transition: 2m }
-    ]
-}
-
-show "Measure Transition Test" {
-    @4/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Measure Transition Test").unwrap();
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-
-        // Verify transition type is Measures
-        assert_eq!(tempo_map.changes.len(), 1);
-        match tempo_map.changes[0].transition {
-            TempoTransition::Measures(m, _) => assert_eq!(m, 2.0),
-            _ => panic!("Expected Measures transition"),
-        }
-
-        // Transition starts at measure 4 (6.0s at 120 BPM)
-        // Transition duration: 2 measures at 4/4 = 8 beats at 120 BPM = 4.0s
-        let change_time = tempo_map.changes[0].position.absolute_time().unwrap();
-
-        // At start of transition: should be 120 BPM
-        let bpm_start = tempo_map.bpm_at_time(change_time, 0.0);
-        assert!((bpm_start - 120.0).abs() < 0.1);
-
-        // During transition (midway): should be interpolated
-        let mid_time = change_time + Duration::from_secs(2); // 2 seconds into 4-second transition
-        let bpm_mid = tempo_map.bpm_at_time(mid_time, 0.0);
-        assert!(
-            (bpm_mid - 150.0).abs() < 1.0,
-            "BPM at transition midpoint should be ~150, got {}",
-            bpm_mid
-        );
-
-        // After transition: should be 180 BPM
-        let end_time = change_time + Duration::from_secs(5); // After transition completes
-        let bpm_end = tempo_map.bpm_at_time(end_time, 0.0);
-        assert!((bpm_end - 180.0).abs() < 0.1);
-    }
-
-    #[test]
-    fn test_end_to_end_multiple_file_level_tempo_sections() {
-        // Test that multiple file-level tempo sections - last one wins
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-tempo {
-    start: 0.0s
-    bpm: 60
-    time_signature: 4/4
-}
-
-show "Multiple Tempo Test" {
-    @1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Multiple Tempo Test").unwrap();
-
-        // Last tempo section should win (60 BPM)
-        assert!(show.tempo_map.is_some());
-        assert_eq!(show.tempo_map.as_ref().unwrap().initial_bpm, 60.0);
-    }
-
-    #[test]
-    fn test_end_to_end_multiple_tempo_sections_in_show() {
-        // Test that multiple tempo sections in one show - last one wins
-        let content = r#"show "Multiple Show Tempo" {
-    tempo {
-        start: 0.0s
-        bpm: 120
-        time_signature: 4/4
-    }
-    
-    tempo {
-        start: 0.0s
-        bpm: 60
-        time_signature: 4/4
-    }
-    
-    @1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Multiple Show Tempo").unwrap();
-
-        // Last tempo section should win (60 BPM)
-        assert!(show.tempo_map.is_some());
-        assert_eq!(show.tempo_map.as_ref().unwrap().initial_bpm, 60.0);
-    }
-
-    #[test]
-    fn test_end_to_end_fractional_measure_duration() {
-        // Test that fractional measure durations convert correctly
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Fractional Measure Duration" {
-    @1/1
-    front_wash: static color: "blue", duration: 1.5measures
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Fractional Measure Duration").unwrap();
-
-        // At 120 BPM in 4/4: 1.5 measures = 6 beats = 3.0s
-        let effect = &show.cues[0].effects[0];
-        let duration = effect.effect_type.get_duration().unwrap();
-        assert!(
-            (duration.as_secs_f64() - 3.0).abs() < 0.001,
-            "1.5 measures should be 3.0s at 120 BPM in 4/4, got {}s",
-            duration.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_end_to_end_consecutive_gradual_transitions() {
-        // Test that consecutive gradual transitions work correctly
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 140, transition: 2 },
-        @6/1 { bpm: 160, transition: 2 }
-    ]
-}
-
-show "Consecutive Transitions" {
-    @4/1
-    front_wash: static color: "blue"
-    
-    @6/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Consecutive Transitions").unwrap();
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-
-        assert_eq!(tempo_map.changes.len(), 2);
-
-        // First transition: 120 -> 140 over 2 beats
-        // Second transition: 140 -> 160 over 2 beats
-        // Verify BPM at various points
-        let change1_time = tempo_map.changes[0].position.absolute_time().unwrap();
-        let change2_time = tempo_map.changes[1].position.absolute_time().unwrap();
-
-        // Before first transition: 120 BPM
-        let bpm_before = tempo_map.bpm_at_time(change1_time - Duration::from_millis(100), 0.0);
-        assert!((bpm_before - 120.0).abs() < 0.1);
-
-        // After first transition completes: 140 BPM
-        let bpm_after1 = tempo_map.bpm_at_time(change1_time + Duration::from_secs(2), 0.0);
-        assert!((bpm_after1 - 140.0).abs() < 1.0);
-
-        // After second transition completes: 160 BPM
-        let bpm_after2 = tempo_map.bpm_at_time(change2_time + Duration::from_secs(2), 0.0);
-        assert!((bpm_after2 - 160.0).abs() < 1.0);
-    }
-
-    #[test]
-    fn test_end_to_end_measure_transition_with_time_signature_change() {
-        // Test measure-based transition when time signature changes during transition
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 140, transition: 2m },
-        @5/1 { time_signature: 3/4 }
-    ]
-}
-
-show "Measure Transition Time Sig Change" {
-    @4/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Measure Transition Time Sig Change").unwrap();
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-
-        // The transition should complete correctly even with time signature change
-        // Transition: 2 measures at 4/4 = 8 beats at 120 BPM = 4.0s
-        let change_time = tempo_map.changes[0].position.absolute_time().unwrap();
-
-        // After transition completes: should be 140 BPM
-        let bpm_after = tempo_map.bpm_at_time(change_time + Duration::from_secs(5), 0.0);
-        assert!((bpm_after - 140.0).abs() < 1.0);
-    }
-
-    #[test]
-    fn test_end_to_end_empty_tempo_section_with_measure_timing() {
-        // Test that empty tempo section works (uses defaults: 120 BPM, 4/4)
-        let content = r#"tempo {
-}
-
-show "Empty Tempo Test" {
-    @1/1
-    front_wash: static color: "blue"
-    
-    @2/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Empty tempo section should work with defaults"
-        );
-        let shows = result.unwrap();
-        let show = shows.get("Empty Tempo Test").unwrap();
-
-        // Should use defaults: 120 BPM, 4/4, start: 0.0s
-        assert!(show.tempo_map.is_some());
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-        assert_eq!(tempo_map.initial_bpm, 120.0);
-        assert_eq!(tempo_map.initial_time_signature.numerator, 4);
-        assert_eq!(tempo_map.initial_time_signature.denominator, 4);
-
-        // At 120 BPM in 4/4: measure 1 = 0.0s, measure 2 = 2.0s
-        assert!((show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001);
-        assert!((show.cues[1].time.as_secs_f64() - 2.0).abs() < 0.001);
-    }
-
-    #[test]
-    fn test_end_to_end_incomplete_tempo_section_with_measure_timing() {
-        // Test that incomplete tempo section (missing bpm or time_signature) still works with defaults
-        let content = r#"tempo {
-    start: 0.0s
-}
-
-show "Incomplete Tempo Test" {
-    @1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Incomplete tempo section should use defaults"
-        );
-        let shows = result.unwrap();
-        let show = shows.get("Incomplete Tempo Test").unwrap();
-
-        // Should use defaults for missing fields
-        assert!(show.tempo_map.is_some());
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-        assert_eq!(tempo_map.initial_bpm, 120.0); // Default
-        assert_eq!(tempo_map.initial_time_signature.numerator, 4); // Default
-        assert_eq!(tempo_map.initial_time_signature.denominator, 4); // Default
-    }
-
-    #[test]
-    fn test_end_to_end_negative_start_offset_rejected() {
-        // Test that negative start offsets are rejected (grammar level)
-        // The grammar uses ASCII_DIGIT+ which doesn't include '-', so it should fail to parse
-        let content = r#"tempo {
-    start: -5.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Negative Start Test" {
-    @1/1
-    front_wash: static color: "blue"
-}"#;
-
-        let result = parse_light_shows(content);
-        // Should fail at grammar level since '-' is not part of ASCII_DIGIT
-        assert!(
-            result.is_err(),
-            "Negative start offset should fail to parse"
-        );
-        if let Err(e) = result {
-            let error_msg = e.to_string();
-            println!("Error message: {}", error_msg);
-            // The error should indicate parsing failure
-            assert!(
-                error_msg.contains("parse")
-                    || error_msg.contains("DSL")
-                    || error_msg.contains("error"),
-                "Error should indicate parsing failure"
-            );
-        }
-    }
-
-    #[test]
-    fn test_end_to_end_very_high_measure_numbers() {
-        // Test that very high measure numbers work correctly
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "High Measures Test" {
-    @1000/1
-    front_wash: static color: "blue"
-    
-    @5000/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "High measure numbers should work");
-        let shows = result.unwrap();
-        let show = shows.get("High Measures Test").unwrap();
-
-        // At 120 BPM in 4/4: measure 1000 = 1998.0s (999 measures * 2s/measure)
-        // At 120 BPM in 4/4: measure 5000 = 9998.0s (4999 measures * 2s/measure)
-        let time1 = show.cues[0].time.as_secs_f64();
-        let time2 = show.cues[1].time.as_secs_f64();
-
-        assert!(
-            time1 > 1990.0 && time1 < 2010.0,
-            "Measure 1000 should be around 1998s, got {}s",
-            time1
-        );
-        assert!(
-            time2 > 9990.0 && time2 < 10010.0,
-            "Measure 5000 should be around 9998s, got {}s",
-            time2
-        );
-        assert!(time2 > time1, "Measure 5000 should be after measure 1000");
-    }
-
-    #[test]
-    fn test_end_to_end_transition_spanning_multiple_changes() {
-        // Test that a gradual transition works correctly even when other changes occur
-        // Use a transition that spans multiple measures, with a change happening after it completes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @4/1 { bpm: 140, transition: 8 },
-        @7/1 { bpm: 160 },
-        @10/1 { time_signature: 3/4 }
-    ]
-}
-
-show "Transition Spanning Changes" {
-    @4/1
-    front_wash: static color: "blue"
-    
-    @10/1
-    back_wash: static color: "red"
-}"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok());
-        let shows = result.unwrap();
-        let show = shows.get("Transition Spanning Changes").unwrap();
-        let tempo_map = show.tempo_map.as_ref().unwrap();
-
-        assert_eq!(tempo_map.changes.len(), 3);
-
-        // First transition: 120 -> 140 over 8 beats at 120 BPM = 4.0s
-        // Transition starts at measure 4 (6.0s) and completes at 10.0s
-        // Second change: snap to 160 at measure 7 (should be after transition completes)
-        // Third change: time signature to 3/4 at measure 10
-        let change1_time = tempo_map.changes[0].position.absolute_time().unwrap();
-        let change2_time = tempo_map.changes[1].position.absolute_time().unwrap();
-
-        // During first transition (early): should be interpolating 120 -> 140
-        let early_time = change1_time + Duration::from_secs(1); // 1 second into 4-second transition
-        let bpm_early = tempo_map.bpm_at_time(early_time, 0.0);
-        // At 25% through transition: 120 + (140-120)*0.25 = 125
-        assert!(
-            (bpm_early - 125.0).abs() < 2.0,
-            "BPM early in transition should be ~125, got {}",
-            bpm_early
-        );
-
-        // During first transition (midway): should be interpolating
-        let mid_time = change1_time + Duration::from_secs(2); // 2 seconds into 4-second transition
-        let bpm_mid = tempo_map.bpm_at_time(mid_time, 0.0);
-        // At 50% through transition: 120 + (140-120)*0.5 = 130
-        assert!(
-            (bpm_mid - 130.0).abs() < 2.0,
-            "BPM at transition midpoint should be ~130, got {}",
-            bpm_mid
-        );
-
-        // After first transition completes but before second change: should be 140
-        // Transition completes at 10.0s, change2 should be after that
-        let after_transition = change1_time + Duration::from_secs(5); // After transition completes
-        let bpm_after_transition = tempo_map.bpm_at_time(after_transition, 0.0);
-        assert!(
-            (bpm_after_transition - 140.0).abs() < 1.0,
-            "BPM after transition completes should be 140, got {}",
-            bpm_after_transition
-        );
-
-        // After second change: should be 160
-        let after_change2 = change2_time + Duration::from_millis(100);
-        let bpm_after2 = tempo_map.bpm_at_time(after_change2, 0.0);
-        assert!((bpm_after2 - 160.0).abs() < 0.1);
-    }
-
-    #[test]
-    fn test_sequence_definition_and_reference() {
-        let content = r#"
-sequence "color_cycle" {
-    @0.000
-    front_wash: static, color: "red"
-    
-    @2.000
-    front_wash: static, color: "green"
-    
-    @4.000
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "color_cycle"
-    
-    @6.000
-    sequence "color_cycle"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse shows with sequences: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        assert_eq!(shows.len(), 1);
-
-        let show = shows.get("Test Show").unwrap();
-        // The sequence should be expanded into 6 cues (3 from first reference, 3 from second)
-        assert_eq!(show.cues.len(), 6);
-
-        // First sequence reference: cues at 0.000, 2.000, 4.000
-        assert_eq!(show.cues[0].time, Duration::from_millis(0));
-        assert_eq!(show.cues[1].time, Duration::from_millis(2000));
-        assert_eq!(show.cues[2].time, Duration::from_millis(4000));
-
-        // Second sequence reference: cues at 6.000, 8.000, 10.000
-        assert_eq!(show.cues[3].time, Duration::from_millis(6000));
-        assert_eq!(show.cues[4].time, Duration::from_millis(8000));
-        assert_eq!(show.cues[5].time, Duration::from_millis(10000));
-    }
-
-    #[test]
-    fn test_sequence_with_effects_in_same_cue() {
-        let content = r#"
-sequence "simple_sequence" {
-    @0.000
-    front_wash: static, color: "red"
-    
-    @1.000
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @5.000
-    back_wash: static, color: "green"
-    sequence "simple_sequence"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse shows: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Should have 3 cues: one with the effect, plus 2 from the sequence
-        // The effect should be added to the first expanded cue
-        assert_eq!(show.cues.len(), 2);
-
-        // First cue: at 5.000 with both the effect and the first sequence cue
-        assert_eq!(show.cues[0].time, Duration::from_millis(5000));
-        assert_eq!(show.cues[0].effects.len(), 2); // green effect + red from sequence
-
-        // Second cue: at 6.000 with the second sequence cue
-        assert_eq!(show.cues[1].time, Duration::from_millis(6000));
-        assert_eq!(show.cues[1].effects.len(), 1); // blue from sequence
-    }
-
-    #[test]
-    fn test_sequence_not_found_error() {
-        let content = r#"
-show "Test Show" {
-    @0.000
-    sequence "nonexistent_sequence"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_err(), "Should fail when sequence is not found");
-
-        let error = result.unwrap_err();
-        assert!(
-            error.to_string().contains("not found"),
-            "Error should mention sequence not found"
-        );
-    }
-
-    #[test]
-    fn test_sequence_with_measure_based_timing() {
-        let content = r#"
-tempo {
-    bpm: 120
-    time_signature: 4/4
-}
-
-sequence "measure_based_sequence" {
-    @1/1
-    front_wash: static, color: "red"
-    
-    @1/3
-    front_wash: static, color: "green"
-    
-    @2/1
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "measure_based_sequence"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse shows with measure-based sequence: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Sequence should be expanded into 3 cues
-        assert_eq!(show.cues.len(), 3);
-
-        // At 120 BPM, 4/4 time:
-        // Measure 1, beat 1 = 0.0s
-        // Measure 1, beat 3 = 1.0s (2 beats at 120 BPM = 1 second)
-        // Measure 2, beat 1 = 2.0s (4 beats at 120 BPM = 2 seconds)
-        // Since the sequence is referenced at 0.000, the times should be offset by 0
-
-        // First cue: measure 1, beat 1 = 0.0s
-        assert_eq!(show.cues[0].time, Duration::from_secs(0));
-
-        // Second cue: measure 1, beat 3 = 1.0s
-        assert_eq!(show.cues[1].time, Duration::from_secs(1));
-
-        // Third cue: measure 2, beat 1 = 2.0s
-        assert_eq!(show.cues[2].time, Duration::from_secs(2));
-    }
-
-    #[test]
-    fn test_sequence_with_own_tempo_and_measure_timing() {
-        let content = r#"
-sequence "sequence_with_tempo" {
-    tempo {
-        bpm: 60
-        time_signature: 4/4
-    }
-    
-    @1/1
-    front_wash: static, color: "red"
-    
-    @2/1
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "sequence_with_tempo"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse shows with sequence having own tempo: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Sequence should be expanded into 2 cues
-        assert_eq!(show.cues.len(), 2);
-
-        // At 60 BPM, 4/4 time:
-        // Measure 1, beat 1 = 0.0s
-        // Measure 2, beat 1 = 4.0s (4 beats at 60 BPM = 4 seconds)
-
-        // First cue: measure 1, beat 1 = 0.0s
-        assert_eq!(show.cues[0].time, Duration::from_secs(0));
-
-        // Second cue: measure 2, beat 1 = 4.0s
-        assert_eq!(show.cues[1].time, Duration::from_secs(4));
-    }
-
-    #[test]
-    fn test_sequence_measure_timing_with_offset() {
-        let content = r#"
-tempo {
-    bpm: 120
-    time_signature: 4/4
-}
-
-sequence "measure_sequence" {
-    @1/1
-    front_wash: static, color: "red"
-    
-    @1/3
-    front_wash: static, color: "green"
-}
-
-show "Test Show" {
-    @10.000
-    sequence "measure_sequence"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Sequence should be expanded with times offset by 10 seconds
-        assert_eq!(show.cues.len(), 2);
-
-        // At 120 BPM:
-        // Measure 1, beat 1 = 0.0s, offset to 10.0s
-        // Measure 1, beat 3 = 1.0s, offset to 11.0s
-
-        assert_eq!(show.cues[0].time, Duration::from_secs(10));
-        assert_eq!(show.cues[1].time, Duration::from_secs(11));
-    }
-
-    #[test]
-    fn test_sequence_looping_finite() {
-        let content = r#"
-sequence "simple_sequence" {
-    @0.000
-    front_wash: static, color: "red"
-    
-    @1.000
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "simple_sequence", loop: 3
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Should have 6 cues (2 cues per iteration × 3 iterations)
-        assert_eq!(show.cues.len(), 6);
-
-        // First iteration: 0s, 1s
-        assert_eq!(show.cues[0].time, Duration::from_secs(0));
-        assert_eq!(show.cues[1].time, Duration::from_secs(1));
-
-        // Second iteration: 1s (last cue time), 2s
-        // Sequence duration is 1s (last cue time since effects are perpetual)
-        assert_eq!(show.cues[2].time, Duration::from_secs(1));
-        assert_eq!(show.cues[3].time, Duration::from_secs(2));
-
-        // Third iteration: 2s, 3s
-        assert_eq!(show.cues[4].time, Duration::from_secs(2));
-        assert_eq!(show.cues[5].time, Duration::from_secs(3));
-
-        // All effects should be marked with sequence name
-        for cue in &show.cues {
-            for effect in &cue.effects {
-                assert_eq!(effect.sequence_name, Some("simple_sequence".to_string()));
-            }
-        }
-    }
-
-    #[test]
-    fn test_sequence_looping_infinite() {
-        let content = r#"
-sequence "infinite_sequence" {
-    @0.000
-    front_wash: static, color: "red"
-    
-    @1.000
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "infinite_sequence", loop: loop
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Should have 20000 cues (2 cues per iteration × 10000 iterations)
-        assert_eq!(show.cues.len(), 20000);
-
-        // Check first few iterations
-        // Sequence duration is 1s (last cue time since effects are perpetual)
-        assert_eq!(show.cues[0].time, Duration::from_secs(0));
-        assert_eq!(show.cues[1].time, Duration::from_secs(1));
-        assert_eq!(show.cues[2].time, Duration::from_secs(1)); // Second iteration starts at 1s
-        assert_eq!(show.cues[3].time, Duration::from_secs(2));
-    }
-
-    #[test]
-    fn test_sequence_looping_once() {
-        let content = r#"
-sequence "once_sequence" {
-    @0.000
-    front_wash: static, color: "red"
-    
-    @1.000
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "once_sequence", loop: once
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Should have 2 cues (default behavior, same as no loop parameter)
-        assert_eq!(show.cues.len(), 2);
-    }
-
-    #[test]
-    fn test_stop_sequence_command() {
-        let content = r#"
-sequence "looping_sequence" {
-    @0.000
-    front_wash: static, color: "red"
-    
-    @1.000
-    front_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "looping_sequence", loop: loop
-    
-    @10.000
-    stop sequence "looping_sequence"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Find the cue at 10 seconds with stop_sequences
-        // Note: The looping sequence may also create a cue at 10 seconds, so we need to find the one with stop_sequences
-        let stop_cue = show
-            .cues
-            .iter()
-            .find(|c| c.time == Duration::from_secs(10) && !c.stop_sequences.is_empty());
-        let cue_times: Vec<_> = show.cues.iter().map(|c| c.time).collect();
-        let cues_at_10: Vec<_> = show
-            .cues
-            .iter()
-            .filter(|c| c.time == Duration::from_secs(10))
-            .map(|c| (c.time, c.stop_sequences.clone(), c.effects.len()))
-            .collect();
-        assert!(stop_cue.is_some(),
-                "Should have a cue at 10 seconds with stop_sequences. Cue times: {:?}, Cues at 10s: {:?}", 
-                cue_times, cues_at_10);
-
-        let stop_cue = stop_cue.unwrap();
-        assert_eq!(
-            stop_cue.stop_sequences,
-            vec!["looping_sequence"],
-            "Stop sequences: {:?}",
-            stop_cue.stop_sequences
-        );
-    }
-
-    #[test]
-    fn test_stop_multiple_sequences() {
-        let content = r#"
-sequence "seq1" {
-    @0.000
-    front_wash: static, color: "red"
-}
-
-sequence "seq2" {
-    @0.000
-    back_wash: static, color: "blue"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "seq1", loop: loop
-    sequence "seq2", loop: loop
-    
-    @5.000
-    stop sequence "seq1"
-    stop sequence "seq2"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // Find the cue at 5 seconds with stop_sequences
-        // Note: The looping sequences may also create cues at 5 seconds, so we need to find the one with stop_sequences
-        let stop_cue = show
-            .cues
-            .iter()
-            .find(|c| c.time == Duration::from_secs(5) && !c.stop_sequences.is_empty());
-        let cue_times: Vec<_> = show.cues.iter().map(|c| c.time).collect();
-        let cues_at_5: Vec<_> = show
-            .cues
-            .iter()
-            .filter(|c| c.time == Duration::from_secs(5))
-            .map(|c| (c.time, c.stop_sequences.clone(), c.effects.len()))
-            .collect();
-        assert!(
-            stop_cue.is_some(),
-            "Should have a cue at 5 seconds with stop_sequences. Cue times: {:?}, Cues at 5s: {:?}",
-            cue_times,
-            cues_at_5
-        );
-
-        let stop_cue = stop_cue.unwrap();
-        assert_eq!(stop_cue.stop_sequences.len(), 2);
-        assert!(stop_cue.stop_sequences.contains(&"seq1".to_string()));
-        assert!(stop_cue.stop_sequences.contains(&"seq2".to_string()));
-    }
-
-    #[test]
-    fn test_nested_sequences() {
-        let content = r#"
-sequence "base_pattern" {
-    @0.000
-    front_wash: static, color: "red"
-    
-    @1.000
-    front_wash: static, color: "blue"
-}
-
-sequence "complex_pattern" {
-    @0.000
-    sequence "base_pattern"
-    back_wash: static, color: "green"
-    
-    @3.000
-    sequence "base_pattern"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "complex_pattern"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse nested sequences: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test Show").unwrap();
-
-        // complex_pattern expands to:
-        // - base_pattern at 0s (2 cues: 0s, 1s) + green effect at 0s (merged into first cue)
-        // - base_pattern at 3s (2 cues: 3s, 4s)
-        // Total: 4 cues (green effect is merged with first base_pattern cue)
-        assert_eq!(show.cues.len(), 4);
-
-        // First base_pattern iteration (with green effect merged)
-        assert_eq!(show.cues[0].time, Duration::from_secs(0));
-        assert_eq!(show.cues[1].time, Duration::from_secs(1));
-
-        // Second base_pattern iteration
-        assert_eq!(show.cues[2].time, Duration::from_secs(3));
-        assert_eq!(show.cues[3].time, Duration::from_secs(4));
-    }
-
-    #[test]
-    fn test_circular_sequence_reference() {
-        // With two-pass parsing, forward references are now supported
-        // The circular reference will be detected when seq_a tries to expand seq_b,
-        // which then tries to expand seq_a (already in recursion stack)
-        let content = r#"
-sequence "seq_a" {
-    @0.000
-    front_wash: static, color: "red"
-    @1.000
-    sequence "seq_b"
-}
-
-sequence "seq_b" {
-    @0.000
-    sequence "seq_a"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "seq_a"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_err(), "Should fail with circular reference");
-
-        let error = result.unwrap_err();
-        assert!(
-            error.to_string().contains("Circular sequence reference"),
-            "Error should mention circular reference: {}",
-            error
-        );
-    }
-
-    #[test]
-    fn test_nested_sequences_measure_timing() {
-        // Test that nested sequences with measure-based timing work correctly
-        // When "verse" is referenced at @17/1, its @1/1 cue should map to @17/1 in the show
-        let content = r#"tempo {
-    start: 3.0s
-    bpm: 160
-    time_signature: 4/4
-}
-
-sequence "verse-start" {
-    @1/1
-    all_wash: static, color: "white"
-}
-
-sequence "verse" {
-    @1/1
-    sequence "verse-start", loop: 1
-    @13/1
-    all_wash: static, color: "red"
-}
-
-show "Test" {
-    @17/1
-    sequence "verse"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse nested sequences with measure timing: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // At 160 BPM in 4/4: 1 beat = 0.375s, 1 measure = 1.5s
-        // Measure 17, beat 1 = 3.0s (start offset) + (16 measures * 1.5s) = 3.0s + 24.0s = 27.0s
-        // "verse-start" should start at measure 17 of the show = 27.0s
-        // "verse" @13/1 should be at measure 17 + 12 measures = measure 29 = 3.0s + (28 * 1.5s) = 3.0s + 42.0s = 45.0s
-
-        // Find the first cue (should be verse-start at 27.0s)
-        // verse-start should start at measure 17 = 27.0s
-        let expected_time = Duration::from_secs_f64(27.0);
-        assert!(!show.cues.is_empty(), "Should have at least one cue");
-        let first_cue_time = show.cues[0].time;
-        assert!(
-            (first_cue_time.as_secs_f64() - expected_time.as_secs_f64()).abs() < 0.001,
-            "verse-start should start at measure 17 (27.0s), got {:?}",
-            first_cue_time
-        );
-    }
-
-    #[test]
-    fn test_self_referencing_sequence() {
-        let content = r#"
-sequence "self_ref" {
-    @0.000
-    sequence "self_ref"
-}
-
-show "Test Show" {
-    @0.000
-    sequence "self_ref"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_err(), "Should fail with self-reference");
-
-        let error = result.unwrap_err();
-        assert!(
-            error.to_string().contains("Circular sequence reference"),
-            "Error should mention circular reference: {}",
-            error
-        );
-    }
-
-    #[test]
-    fn test_measure_offset() {
-        // Test that offset commands correctly adjust measure numbers
-        // Measures 1-8 repeat once, so measure 9 should be at playback measure 17
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Test" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    offset 8 measures
-    
-    @1/1
-    all_wash: static, color: "green"
-    
-    @8/1
-    all_wash: static, color: "yellow"
-    
-    @9/1
-    all_wash: static, color: "red"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with offset: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // At 120 BPM in 4/4: 1 measure = 2.0s
-        // Measure 1, beat 1 = 0.0s + (0 measures * 2.0s) = 0.0s
-        // After offset 8: measure 1 becomes playback measure 9 = 0.0s + (8 measures * 2.0s) = 16.0s
-        // After offset 8: measure 8 becomes playback measure 16 = 0.0s + (15 measures * 2.0s) = 30.0s
-        // After offset 8: measure 9 becomes playback measure 17 = 0.0s + (16 measures * 2.0s) = 32.0s
-
-        assert!(show.cues.len() >= 4, "Should have at least 4 cues");
-
-        // First cue at measure 1 = 0.0s
-        let first_cue_time = show.cues[0].time;
-        let expected_first = Duration::from_secs_f64(0.0);
-        assert!(
-            (first_cue_time.as_secs_f64() - expected_first.as_secs_f64()).abs() < 0.001,
-            "First cue should be at measure 1 (0.0s), got {:?}",
-            first_cue_time
-        );
-
-        // Second cue at measure 1 (with offset) = playback measure 9 = 16.0s
-        let second_cue_time = show.cues[1].time;
-        let expected_second = Duration::from_secs_f64(16.0);
-        assert!(
-            (second_cue_time.as_secs_f64() - expected_second.as_secs_f64()).abs() < 0.001,
-            "Second cue should be at playback measure 9 (16.0s), got {:?}",
-            second_cue_time
-        );
-
-        // Third cue at measure 8 (with offset) = playback measure 16 = 30.0s
-        let third_cue_time = show.cues[2].time;
-        let expected_third = Duration::from_secs_f64(30.0);
-        assert!(
-            (third_cue_time.as_secs_f64() - expected_third.as_secs_f64()).abs() < 0.001,
-            "Third cue should be at playback measure 16 (30.0s), got {:?}",
-            third_cue_time
-        );
-
-        // Fourth cue at measure 9 (with offset) = playback measure 17 = 32.0s
-        let fourth_cue_time = show.cues[3].time;
-        let expected_fourth = Duration::from_secs_f64(32.0);
-        assert!(
-            (fourth_cue_time.as_secs_f64() - expected_fourth.as_secs_f64()).abs() < 0.001,
-            "Fourth cue should be at playback measure 17 (32.0s), got {:?}",
-            fourth_cue_time
-        );
-    }
-
-    #[test]
-    fn test_reset_measures() {
-        // Test that reset_measures resets the offset back to 0
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Test" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    offset 8 measures
-    
-    @1/1
-    all_wash: static, color: "green"
-    
-    reset_measures
-    
-    @1/1
-    all_wash: static, color: "red"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with reset_measures: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // At 120 BPM in 4/4: 1 measure = 2.0s
-        // First cue at measure 1 = 0.0s
-        // Second cue at measure 1 (with offset 8) = playback measure 9 = 16.0s
-        // Third cue at measure 1 (after reset) = 0.0s again
-
-        assert!(show.cues.len() >= 3, "Should have at least 3 cues");
-
-        let first_cue_time = show.cues[0].time;
-        let second_cue_time = show.cues[1].time;
-        let third_cue_time = show.cues[2].time;
-
-        assert!(
-            (first_cue_time.as_secs_f64() - 0.0).abs() < 0.001,
-            "First cue should be at 0.0s, got {:?}",
-            first_cue_time
-        );
-
-        assert!(
-            (second_cue_time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Second cue should be at 16.0s (measure 9), got {:?}",
-            second_cue_time
-        );
-
-        assert!(
-            (third_cue_time.as_secs_f64() - 0.0).abs() < 0.001,
-            "Third cue should be at 0.0s (after reset), got {:?}",
-            third_cue_time
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_accumulation() {
-        // Test that multiple offset commands accumulate
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Test" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    offset 4 measures
-    
-    @1/1
-    all_wash: static, color: "green"
-    
-    offset 4 measures
-    
-    @1/1
-    all_wash: static, color: "red"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with accumulating offsets: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // At 120 BPM in 4/4: 1 measure = 2.0s
-        // First cue at measure 1 = 0.0s
-        // After offset 4: measure 1 becomes playback measure 5 = 8.0s
-        // After offset 8 (4+4): measure 1 becomes playback measure 9 = 16.0s
-
-        assert!(show.cues.len() >= 3, "Should have at least 3 cues");
-
-        assert!(
-            (show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001,
-            "First cue should be at 0.0s"
-        );
-
-        assert!(
-            (show.cues[1].time.as_secs_f64() - 8.0).abs() < 0.001,
-            "Second cue should be at 8.0s (measure 5)"
-        );
-
-        assert!(
-            (show.cues[2].time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Third cue should be at 16.0s (measure 9)"
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_in_sequence() {
-        // Test that offset works correctly in sequences
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-sequence "verse" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    offset 8 measures
-    
-    @1/1
-    all_wash: static, color: "green"
-    
-    @9/1
-    all_wash: static, color: "red"
-}
-
-show "Test" {
-    @1/1
-    sequence "verse"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse sequence with offset: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // Sequence is referenced at measure 1, so its @1/1 becomes measure 1
-        // After offset 8 in sequence: @1/1 becomes playback measure 9
-        // After offset 8 in sequence: @9/1 becomes playback measure 17
-        // But these are relative to sequence start (measure 1), so:
-        // First cue: measure 1 = 0.0s
-        // Second cue: measure 9 = 16.0s
-        // Third cue: measure 17 = 32.0s
-
-        assert!(show.cues.len() >= 3, "Should have at least 3 cues");
-
-        assert!(
-            (show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001,
-            "First cue should be at 0.0s"
-        );
-
-        assert!(
-            (show.cues[1].time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Second cue should be at 16.0s (measure 9)"
-        );
-
-        assert!(
-            (show.cues[2].time.as_secs_f64() - 32.0).abs() < 0.001,
-            "Third cue should be at 32.0s (measure 17)"
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_with_fractional_beats() {
-        // Test that offset works with fractional beats
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Test" {
-    @1/1.5
-    all_wash: static, color: "blue"
-    
-    offset 8 measures
-    
-    @1/2.5
-    all_wash: static, color: "green"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with offset and fractional beats: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // At 120 BPM in 4/4: 1 beat = 0.5s
-        // First cue: measure 1, beat 1.5 = 0.0s + (0 measures + 0.5 beats) * 0.5s = 0.25s
-        // After offset 8: measure 1, beat 2.5 = (8 measures + 1.5 beats) * 0.5s = 16.0s + 0.75s = 16.75s
-
-        assert!(show.cues.len() >= 2, "Should have at least 2 cues");
-
-        assert!(
-            (show.cues[0].time.as_secs_f64() - 0.25).abs() < 0.001,
-            "First cue should be at 0.25s"
-        );
-
-        assert!(
-            (show.cues[1].time.as_secs_f64() - 16.75).abs() < 0.001,
-            "Second cue should be at 16.75s"
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_with_tempo_changes() {
-        // Test that offset works correctly with tempo changes
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-    changes: [
-        @8/1 {
-            bpm: 60
-        }
-    ]
-}
-
-show "Test" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    offset 8 measures
-    
-    @1/1
-    all_wash: static, color: "green"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with offset and tempo change: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // At 120 BPM in 4/4: 1 measure = 2.0s
-        // With offset 8 applied to tempo changes:
-        // - Tempo change at score measure 8/1 + offset 8 = playback measure 16
-        // - Second cue at score measure 1 + offset 8 = playback measure 9
-        // Since the tempo change is at playback measure 16 (after the cue at measure 9),
-        // the second cue is still at 120 BPM
-        // Measures 1-8 at 120 BPM = 8 * 2.0s = 16.0s
-        // Start of measure 9: 16.0s
-
-        assert!(show.cues.len() >= 2, "Should have at least 2 cues");
-
-        assert!(
-            (show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001,
-            "First cue should be at 0.0s"
-        );
-
-        assert!(
-            (show.cues[1].time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Second cue should be at 16.0s (measure 9 at 120 BPM, tempo change is at measure 16)"
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_in_same_cue() {
-        // Test that when a cue has both a measure time and an offset command,
-        // the offset does NOT apply to that cue's measure time, but DOES apply to subsequent cues.
-        // This tests the scenario where offset is in the same cue definition as a measure time.
-        let content = r#"tempo {
-            start: 1.5s
-            bpm: 160
-            time_signature: 4/4
-            changes: [
-                @68/1 { bpm: 180 },
-                @104/1 { bpm: 160 }
-            ]
-        }
-
-show "Test" {
-    @70/1
-    all_wash: static, color: "blue"
-    
-    @74/1
-    all_wash: static, color: "green"
-    offset 5 measures
-    
-    @70/1
-    all_wash: static, color: "red"
-    
-    @74/1
-    all_wash: static, color: "yellow"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with offset in same cue: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // Verify we have 4 cues
-        assert!(show.cues.len() >= 4, "Should have at least 4 cues");
-
-        // First cue at @70/1 (no offset yet)
-        // Tempo changes at @68/1 to 180 BPM
-        // Measures 1-67 at 160 BPM: 67 * 1.5s = 100.5s
-        // Measures 68-70 at 180 BPM: 2 * 1.333s = 2.667s
-        // Total: 100.5 + 2.667 = 103.167s
-        // Plus start_offset 1.5s: 103.167 + 1.5 = 104.667s
-        let first_cue_time = show.cues[0].time;
-        let expected_first = Duration::from_secs_f64(104.666666667);
-        assert!(
-            (first_cue_time.as_secs_f64() - expected_first.as_secs_f64()).abs() < 0.01,
-            "First cue should be at measure 70 (104.667s), got {:?} ({:.3}s)",
-            first_cue_time,
-            first_cue_time.as_secs_f64()
-        );
-
-        // Second cue at @74/1 (no offset yet)
-        // Measures 1-67 @160 BPM: 100.5s
-        // Measures 68-73 @180 BPM: 6 * 1.333s = 8.0s
-        // Plus start_offset 1.5s => 110.0s
-        let second_cue_time = show.cues[1].time;
-        let expected_second = Duration::from_secs_f64(110.0);
-        assert!(
-            (second_cue_time.as_secs_f64() - expected_second.as_secs_f64()).abs() < 0.01,
-            "Second cue should be at measure 74 (110.0s), got {:?} ({:.3}s)",
-            second_cue_time,
-            second_cue_time.as_secs_f64()
-        );
-
-        // Third cue at @70/1 with offset 5 measures = playback measure 75
-        // Base @70/1 time: 104.667s, offset 5 measures at 180 BPM = 6.667s
-        let third_cue_time = show.cues[2].time;
-        let expected_third = Duration::from_secs_f64(111.333333333);
-        assert!(
-            (third_cue_time.as_secs_f64() - expected_third.as_secs_f64()).abs() < 0.01,
-            "Third cue should be at playback measure 75 (~111.33s) after offset 5, got {:?} ({:.3}s)",
-            third_cue_time,
-            third_cue_time.as_secs_f64()
-        );
-
-        // Fourth cue at @74/1 with offset 5 measures = playback measure 79
-        // Base @74/1 time: 110.0s, offset still 5 measures at 180 BPM = 6.667s => ~116.67s
-        let fourth_cue_time = show.cues[3].time;
-        let expected_fourth = Duration::from_secs_f64(116.666666667);
-        assert!(
-            (fourth_cue_time.as_secs_f64() - expected_fourth.as_secs_f64()).abs() < 0.01,
-            "Fourth cue should be at playback measure 79 (~116.67s) after offset 5, got {:?} ({:.3}s)",
-            fourth_cue_time,
-            fourth_cue_time.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_with_measure_time_in_same_cue() {
-        // Test that when a cue has both a measure time AND an offset command in the same
-        // cue definition, the offset does NOT apply to the current cue's measure time.
-        // The offset should only affect subsequent cues. The effect should still be included.
-        let content = r#"tempo {
-            start: 0.0s
-            bpm: 120
-            time_signature: 4/4
-        }
-
-show "Test" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    @5/1
-    all_wash: static, color: "green"
-    offset 10 measures
-    
-    @1/1
-    all_wash: static, color: "red"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with offset and measure time in same cue: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // Verify we have 3 cues
-        assert!(show.cues.len() >= 3, "Should have at least 3 cues");
-
-        // At 120 BPM in 4/4: 1 measure = 2.0s
-
-        // First cue at @1/1 (no offset) = 0.0s
-        let first_cue_time = show.cues[0].time;
-        let expected_first = Duration::from_secs_f64(0.0);
-        assert!(
-            (first_cue_time.as_secs_f64() - expected_first.as_secs_f64()).abs() < 0.001,
-            "First cue should be at measure 1 (0.0s), got {:?} ({:.3}s)",
-            first_cue_time,
-            first_cue_time.as_secs_f64()
-        );
-        // Verify first cue has the effect
-        assert!(
-            !show.cues[0].effects.is_empty(),
-            "First cue should have effects"
-        );
-
-        // Second cue at @5/1 with offset 10 measures in the SAME cue
-        // IMPORTANT: The offset should NOT apply to this cue's measure time.
-        // So @5/1 should be at playback measure 5 = 4 measures * 2.0s = 8.0s
-        let second_cue_time = show.cues[1].time;
-        let expected_second = Duration::from_secs_f64(8.0);
-        assert!(
-            (second_cue_time.as_secs_f64() - expected_second.as_secs_f64()).abs() < 0.001,
-            "Second cue should be at playback measure 5 (8.0s), offset should NOT apply to current cue, got {:?} ({:.3}s)",
-            second_cue_time,
-            second_cue_time.as_secs_f64()
-        );
-        // Verify second cue has the effect (this is the key test - effect should not be lost)
-        assert!(
-            !show.cues[1].effects.is_empty(),
-            "Second cue should have effects even though it has an offset command"
-        );
-
-        // Third cue at @1/1 (with offset 10 from previous cue) = playback measure 11 = 10 measures * 2.0s = 20.0s
-        let third_cue_time = show.cues[2].time;
-        let expected_third = Duration::from_secs_f64(20.0);
-        assert!(
-            (third_cue_time.as_secs_f64() - expected_third.as_secs_f64()).abs() < 0.001,
-            "Third cue should be at playback measure 11 (20.0s) with offset 10, got {:?} ({:.3}s)",
-            third_cue_time,
-            third_cue_time.as_secs_f64()
-        );
-    }
-
-    #[test]
-    fn test_offset_timing_at_180_bpm_with_tempo_change() {
-        // Test the exact scenario: @70/1, @74/1 with offset 5, @70/1 at 180 BPM
-        // This verifies that the offset is applied correctly and timing is accurate
-        // when there's a tempo change from 160 to 180 BPM at @68/1
-        let content = r#"tempo {
-            start: 1.5s
-            bpm: 160
-            time_signature: 4/4
-            changes: [
-                @68/1 { bpm: 180 },
-                @104/1 { bpm: 160 }
-            ]
-        }
-
-show "Test" {
-    @70/1
-    all_wash: static, color: "blue"
-    
-    @74/1
-    all_wash: static, color: "green"
-    offset 5 measures
-    
-    @70/1
-    all_wash: static, color: "red"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(result.is_ok(), "Failed to parse show: {:?}", result.err());
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        assert!(show.cues.len() >= 3, "Should have at least 3 cues");
-
-        // At 180 BPM in 4/4: 1 measure = 4 beats * (60/180) = 1.333... seconds
-        // Measure 74 to measure 75 = 1 measure = 1.333 seconds
-
-        let second_cue_time = show.cues[1].time; // @74/1
-        let third_cue_time = show.cues[2].time; // @70/1 with offset 5 (should be measure 75)
-
-        println!(
-            "Second cue (@74/1) time: {:.3}s",
-            second_cue_time.as_secs_f64()
-        );
-        println!(
-            "Third cue (@70/1 with offset 5) time: {:.3}s",
-            third_cue_time.as_secs_f64()
-        );
-
-        // Also compute directly from tempo map for visibility
-        let tm = show.tempo_map.as_ref().unwrap();
-        let calc_74 = tm
-            .measure_to_time_with_offset(74, 1.0, 0, 0.0)
-            .unwrap()
-            .as_secs_f64();
-        let calc_70_off5 = tm
-            .measure_to_time_with_offset(70, 1.0, 5, 0.0)
-            .unwrap()
-            .as_secs_f64();
-        println!(
-            "Calc tempo map: @74/1 = {:.3}s, @70/1 (offset 5) = {:.3}s, diff = {:.3}s",
-            calc_74,
-            calc_70_off5,
-            calc_70_off5 - calc_74
-        );
-
-        let time_diff = third_cue_time.as_secs_f64() - second_cue_time.as_secs_f64();
-        let expected_diff = 1.333333333; // 1 measure at 180 BPM
-
-        // Calculate what measure the third cue is actually at based on the time difference
-        let actual_measures = time_diff / 1.333333333; // measures at 180 BPM
-        println!(
-            "Time difference: {:.3}s = {:.3} measures at 180 BPM (expected: 1.0 measure)",
-            time_diff, actual_measures
-        );
-
-        assert!(
-            (time_diff - expected_diff).abs() < 0.01,
-            "Time difference between @74/1 and second @70/1 (with offset 5) should be ~1.333s (1 measure at 180 BPM), got {:.3}s (difference: {:.3}s, actual: {:.3} measures)",
-            time_diff,
-            time_diff - expected_diff,
-            actual_measures
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_at_start() {
-        // Test that offset can be in the first cue
-        let content = r#"tempo {
-            start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Test" {
-    @1/1
-    offset 8 measures
-    
-    @1/1
-    all_wash: static, color: "blue"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with offset in first cue: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // First cue at @1/1 = 0.0s (parsed before offset command), then offset is set to 8
-        // Second cue at @1/1 with offset 8 = playback measure 9 = 16.0s
-
-        assert!(show.cues.len() >= 2, "Should have at least 2 cues");
-
-        assert!(
-            (show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001,
-            "First cue should be at 0.0s (before offset takes effect)"
-        );
-
-        assert!(
-            (show.cues[1].time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Second cue should be at 16.0s (measure 9, after offset)"
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_reset_and_reoffset() {
-        // Test reset followed by another offset
-        // Note: offset commands affect subsequent cues, not the current one
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Test" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    offset 8 measures
-    
-    @1/1
-    all_wash: static, color: "green"
-    
-    @1/1
-    all_wash: static, color: "yellow"
-    reset_measures
-    offset 4 measures
-    
-    @1/1
-    all_wash: static, color: "red"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with reset and reoffset: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // First cue: measure 1 = 0.0s, then offset becomes 8
-        // Second cue: measure 1 with offset 8 = measure 9 = 16.0s
-        // Third cue: measure 1 with offset 8 (still) = measure 9 = 16.0s, then reset and offset 4
-        // Fourth cue: measure 1 with offset 4 (after reset) = measure 5 = 8.0s
-
-        assert!(show.cues.len() >= 4, "Should have at least 4 cues");
-
-        assert!(
-            (show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001,
-            "First cue should be at 0.0s"
-        );
-
-        assert!(
-            (show.cues[1].time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Second cue should be at 16.0s (measure 9)"
-        );
-
-        assert!(
-            (show.cues[2].time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Third cue should be at 16.0s (measure 9, before reset takes effect)"
-        );
-
-        let fourth_cue_time = show.cues[3].time.as_secs_f64();
-        assert!(
-            (fourth_cue_time - 8.0).abs() < 0.001,
-            "Fourth cue should be at 8.0s (measure 5 after reset and offset 4), got {}",
-            fourth_cue_time
-        );
-    }
-
-    #[test]
-    fn test_measure_offset_with_alternate_endings_scenario() {
-        // Test the original use case: measures 1-8 repeat twice, measure 9 is alternate ending
-        // This simulates: measures 1-8 play, then 1-8 again, but measure 9 replaces measure 8 on 2nd repeat
-        let content = r#"tempo {
-    start: 0.0s
-    bpm: 120
-    time_signature: 4/4
-}
-
-show "Test" {
-    @1/1
-    all_wash: static, color: "blue"
-    
-    @8/1
-    all_wash: static, color: "yellow"
-    
-    offset 8 measures
-    
-    @1/1
-    all_wash: static, color: "green"
-    
-    @7/1
-    all_wash: static, color: "orange"
-    
-    @9/1
-    all_wash: static, color: "red"
-    
-    @10/1
-    all_wash: static, color: "purple"
-}
-"#;
-
-        let result = parse_light_shows(content);
-        assert!(
-            result.is_ok(),
-            "Failed to parse show with alternate endings scenario: {:?}",
-            result.err()
-        );
-
-        let shows = result.unwrap();
-        let show = shows.get("Test").unwrap();
-
-        // At 120 BPM in 4/4: 1 measure = 2.0s
-        // First playthrough: measures 1-8
-        //   @1/1 = 0.0s (measure 1)
-        //   @8/1 = 14.0s (measure 8)
-        // Second playthrough (after offset 8): measures 1-8 again, but measure 9 replaces 8
-        //   @1/1 = 16.0s (playback measure 9)
-        //   @7/1 = 28.0s (playback measure 15 = 14 measures * 2.0s)
-        //   @9/1 = 32.0s (playback measure 17 = 16 measures * 2.0s) - this is the alternate ending
-        //   @10/1 = 34.0s (playback measure 18 = 17 measures * 2.0s) - continues after repeat
-
-        assert!(show.cues.len() >= 6, "Should have at least 6 cues");
-
-        assert!(
-            (show.cues[0].time.as_secs_f64() - 0.0).abs() < 0.001,
-            "First cue should be at 0.0s"
-        );
-
-        assert!(
-            (show.cues[1].time.as_secs_f64() - 14.0).abs() < 0.001,
-            "Second cue should be at 14.0s (measure 8)"
-        );
-
-        assert!(
-            (show.cues[2].time.as_secs_f64() - 16.0).abs() < 0.001,
-            "Third cue should be at 16.0s (measure 9, second repeat)"
-        );
-
-        assert!(
-            (show.cues[3].time.as_secs_f64() - 28.0).abs() < 0.001,
-            "Fourth cue should be at 28.0s (measure 15)"
-        );
-
-        assert!(
-            (show.cues[4].time.as_secs_f64() - 32.0).abs() < 0.001,
-            "Fifth cue should be at 32.0s (measure 17, alternate ending)"
-        );
-
-        assert!(
-            (show.cues[5].time.as_secs_f64() - 34.0).abs() < 0.001,
-            "Sixth cue should be at 34.0s (measure 18, after repeat)"
-        );
-    }
-}
-
-#[test]
-fn test_parse_chase_random_pattern() {
-    use crate::lighting::effects::{ChasePattern, EffectType};
-    use crate::lighting::parser::parse_light_shows;
-
-    let content = r#"tempo {
-    start: 1.5s
-    bpm: 160
-    time_signature: 4/4
-    changes: [
-        @68/1 {
-            bpm: 180
-        }
-    ]
-}
-
-show "Test" {
-    @70/1
-    all_wash: chase, speed: 1beats, pattern: random, hold_time: 4measures
-}
-"#;
-
-    let result = parse_light_shows(content);
-    assert!(result.is_ok(), "Failed to parse show: {:?}", result.err());
-
-    let shows = result.unwrap();
-    let show = shows.get("Test").unwrap();
-
-    assert_eq!(show.cues.len(), 1);
-    let cue = &show.cues[0];
-    assert_eq!(cue.effects.len(), 1);
-
-    let effect = &cue.effects[0];
-    match &effect.effect_type {
-        EffectType::Chase {
-            pattern,
-            speed: _,
-            direction: _,
-            transition: _,
-        } => {
-            println!("Parsed pattern: {:?}", pattern);
-            println!("Pattern debug: pattern={:?}", pattern);
-            match pattern {
-                ChasePattern::Random => {
-                    println!("✓ Pattern is correctly set to Random");
-                }
-                ChasePattern::Linear => {
-                    println!("✗ Pattern is Linear, but should be Random!");
-                    panic!("Pattern was parsed as Linear instead of Random");
-                }
-                ChasePattern::Snake => {
-                    println!("✗ Pattern is Snake, but should be Random!");
-                    panic!("Pattern was parsed as Snake instead of Random");
-                }
-            }
-        }
-        other => {
-            println!("Effect type: {:?}", other);
-            panic!("Effect is not a Chase effect! Got: {:?}", other);
-        }
-    }
-}
+mod basic_parsing_tests;
+#[cfg(test)]
+mod chase_pattern_tests;
+#[cfg(test)]
+mod fixture_venue_tests;
+#[cfg(test)]
+mod layer_commands_tests;
+#[cfg(test)]
+mod measure_offset_tests;
+#[cfg(test)]
+mod parameters_tests;
+#[cfg(test)]
+mod sequences_tests;
+#[cfg(test)]
+mod tempo_durations_tests;
+#[cfg(test)]
+mod tempo_end_to_end_tests;
+#[cfg(test)]
+mod tempo_validation_tests;
